Convert a tidygraph to a monarch_kg graph.

Converts a tidygraph of the appropriate format into a monarch_kg graph. Not intended for direct use.

as.monarch_kg(graph_obj, ...)

Arguments

graph_obj

input to convert to a monarch_kg.

...

additional arguments to be passed to or from methods (ignored).

Value

a new monarch_kg.

Details

The monarch_kg class extends tidygraph and defines Monarch-specific methods in addition to functionality provided by tidygraph and igraph.

See also

Examples

g <- monarch_search("Cystic Fibrosis", limit = 3)
#> 
#> 
#> NOTE: Only 3 of 16 results returned.

str(g)
#> Classes 'monarch_kg', 'monarch_kg', 'tbl_graph', 'igraph'  hidden list of 10
#>  $ : num 3
#>  $ : logi TRUE
#>  $ : num(0) 
#>  $ : num(0) 
#>  $ : num(0) 
#>  $ : num(0) 
#>  $ : num [1:4] 0 0 0 0
#>  $ : num [1:4] 0 0 0 0
#>  $ :List of 4
#>   ..$ : num [1:3] 1 0 1
#>   ..$ : Named list()
#>   ..$ :List of 10
#>   .. ..$ id              : chr [1:3] "MONDO:0009061" "MONDO:0005413" "MONDO:0009062"
#>   .. ..$ categories      :List of 3
#>   .. .. ..$ : chr "biolink:Disease"
#>   .. .. ..$ : chr "biolink:Disease"
#>   .. .. ..$ : chr "biolink:Disease"
#>   .. ..$ label           : chr [1:3] "cystic fibrosis" "cystic fibrosis associated meconium ileus" "cystic fibrosis-gastritis-megaloblastic anemia syndrome"
#>   .. ..$ description     : chr [1:3] "Cystic fibrosis (CF) is a genetic disorder characterized by the production of sweat with a high salt content an"| __truncated__ "Cystic fibrosis associated meconium ileum is a thickening and congestion of the meconium in the ileum in newbor"| __truncated__ "A rare genetic disease reported in two siblings of consanguineous Arab parents and is characterized by cystic f"| __truncated__
#>   .. ..$ symbol          : chr [1:3] NA NA NA
#>   .. ..$ synonyms        :List of 3
#>   .. .. ..$ : chr(0) 
#>   .. .. ..$ : chr(0) 
#>   .. .. ..$ : chr(0) 
#>   .. ..$ taxon           : chr [1:3] NA NA NA
#>   .. ..$ inheritance_id  : chr [1:3] "HP:0000007" NA NA
#>   .. ..$ inheritance_name: chr [1:3] "HP:0000007" NA NA
#>   .. ..$ selected        : logi [1:3] TRUE TRUE TRUE
#>   ..$ : Named list()
#>  $ :<environment: 0x7fa85013a0f8> 
#>  - attr(*, "active")= chr "nodes"

print(methods(class = "monarch_kg"))
#> [1] select_nodes
#> see '?methods' for accessing help and source code