Get a list of results matching a search term.
Source: R/monarchr.R

https://api.monarchinitiative.org/api/search/entity/uveal

biolink_search(phrase_or_id, rows = 20)

Arguments

phrase_or_id
Something for which we are looking for more information, or to find useful entities.

Number
of rows of results to fetch.

Value

A list of (the search results, and monarch_api S3 class.)

Details

Note, as opposed to the bioentity_ functions, the biolink_search results are not standardized to a tibble?

https://api.monarchinitiative.org/api/search/entity/autocomplete/MELANOMA

Examples

term <- "uveal melanoma"
biolink_search(term)

#> $search_results
#> # A tibble: 20 x 13
#>    definition   category  label prefix equivalent_iri id    score edges synonym
#>    <chr>        <chr>     <chr> <chr>  <chr>          <chr> <chr> <chr> <chr>  
#>  1 A melanoma … disease   uvea… MONDO  http://purl.o… MOND… 116.… 355   uvea m…
#>  2 A melanoma … disease   mixe… MONDO  http://purl.o… MOND… 105.… 88    mixed …
#>  3 <NA>         disease   post… MONDO  http://purl.o… MOND… 105.… 15    medium…
#>  4 A uveal mel… disease   necr… MONDO  http://purl.o… MOND… 105.… 46    <NA>   
#>  5 A uveal mel… disease   epit… MONDO  http://purl.o… MOND… 105.… 77    uvea e…
#>  6 <NA>         disease   susc… MONDO  http://purl.o… MOND… 105.… 105   MELANO…
#>  7 Uveal melan… disease   inte… MONDO  http://purl.o… MOND… 97.7… 52    intrao…
#>  8 A malignant… Phenotype Intr… HP     _:8bb7af2439a… HP:0… 66.6… 65    Uveal …
#>  9 A non-neopl… disease   uvea… MONDO  http://purl.o… MOND… 62.4… 141   uvea d…
#> 10 greater tha… Phenotype incr… MP     _:2d67f2eeaa8… MP:0… 58.4… 19    increa…
#> 11 <NA>         disease   chor… MONDO  _:7a7ac34b7b8… MOND… 58.4… 52    optic …
#> 12 A uveal mel… disease   iris… MONDO  http://purl.o… MOND… 57.5… 102   malign…
#> 13 A melanoma … disease   spin… MONDO  http://purl.o… MOND… 56.1… 38    Spindl…
#> 14 A rare uvea… disease   mali… MONDO  http://purl.o… MOND… 53.4… 105   Melano…
#> 15 <NA>         disease   cili… MONDO  http://purl.o… MOND… 53.4… 52    Ciliar…
#> 16 Choroid mel… disease   inte… MONDO  http://purl.o… MOND… 53.4… 44    mixed …
#> 17 Iris melano… disease   inte… MONDO  http://purl.o… MOND… 53.4… 60    iris i…
#> 18 <NA>         disease   mela… MONDO  http://purl.o… MOND… 52.7… 35    MELANO…
#> 19 <NA>         disease   mela… MONDO  http://purl.o… MOND… 52.7… 36    MELANO…
#> 20 Ciliary bod… disease   inte… MONDO  http://purl.o… MOND… 51.1… 48    ciliar…
#> # ... with 4 more variables: `_version_` <chr>, equivalent_curie <chr>,
#> #   leaf <chr>, iri <chr>
#> 
#> $response
#> <monarch https://api.monarchinitiative.org/api/search/entity/uveal%20melanoma?rows=20&format=json>
#> <Showing parsed R objects from json response. 2 level(s) deep>
#> List of 4
#>  $ highlighting:List of 20
#>   ..$ MONDO:0003913:List of 3
#>   ..$ MONDO:0004365:List of 3
#>   ..$ MONDO:0003910:List of 3
#>   ..$ MONDO:0003911:List of 3
#>   ..$ MONDO:0004063:List of 3
#>   ..$ MONDO:0004065:List of 3
#>   ..$ MONDO:0003744:List of 3
#>   ..$ MP:0010276   :List of 3
#>   ..$ MONDO:0006200:List of 3
#>   ..$ MONDO:0003927:List of 3
#>   ..$ MONDO:0003912:List of 3
#>   ..$ MONDO:0006486:List of 3
#>   ..$ HP:0007716   :List of 3
#>   ..$ MONDO:0004064:List of 3
#>   ..$ MONDO:0004066:List of 3
#>   ..$ MONDO:0011695:List of 3
#>   ..$ MONDO:0004062:List of 3
#>   ..$ MONDO:0011696:List of 3
#>   ..$ MONDO:0007966:List of 3
#>   ..$ MONDO:0002661:List of 3
#>  $ numFound    : int 704
#>  $ facet_counts:List of 2
#>   ..$ category   :List of 3
#>   ..$ taxon_label:List of 16
#>  $ docs        :'data.frame':	20 obs. of  37 variables:
#>   ..$ definition          :List of 20
#>   ..$ synonym_eng         :List of 20
#>   ..$ synonym_kw          :List of 20
#>   ..$ category_std        :List of 20
#>   ..$ id_eng              : chr [1:20] "MONDO:0006486" "MONDO:0003910" "MONDO:0003927" "MONDO:0004365" ...
#>   ..$ equivalent_iri_std  :List of 20
#>   ..$ equivalent_curie_kw :List of 20
#>   ..$ id_std              : chr [1:20] "MONDO:0006486" "MONDO:0003910" "MONDO:0003927" "MONDO:0004365" ...
#>   ..$ label_std           :List of 20
#>   ..$ definition_std      :List of 20
#>   ..$ label_eng           :List of 20
#>   ..$ category            :List of 20
#>   ..$ label               :List of 20
#>   ..$ category_kw         :List of 20
#>   ..$ prefix              : chr [1:20] "MONDO" "MONDO" "MONDO" "MONDO" ...
#>   ..$ label_kw            :List of 20
#>   ..$ equivalent_iri      :List of 20
#>   ..$ id                  : chr [1:20] "MONDO:0006486" "MONDO:0003910" "MONDO:0003927" "MONDO:0004365" ...
#>   ..$ iri_kw              : chr [1:20] "http://purl.obolibrary.org/obo/MONDO_0006486" "http://purl.obolibrary.org/obo/MONDO_0003910" "http://purl.obolibrary.org/obo/MONDO_0003927" "http://purl.obolibrary.org/obo/MONDO_0004365" ...
#>   ..$ definition_kw       :List of 20
#>   ..$ iri_std             : chr [1:20] "http://purl.obolibrary.org/obo/MONDO_0006486" "http://purl.obolibrary.org/obo/MONDO_0003910" "http://purl.obolibrary.org/obo/MONDO_0003927" "http://purl.obolibrary.org/obo/MONDO_0004365" ...
#>   ..$ category_eng        :List of 20
#>   ..$ equivalent_curie_eng:List of 20
#>   ..$ score               : num [1:20] 116 105 105 105 105 ...
#>   ..$ edges               : int [1:20] 355 88 15 46 77 105 52 65 141 19 ...
#>   ..$ synonym_std         :List of 20
#>   ..$ definition_eng      :List of 20
#>   ..$ iri_eng             : chr [1:20] "http://purl.obolibrary.org/obo/MONDO_0006486" "http://purl.obolibrary.org/obo/MONDO_0003910" "http://purl.obolibrary.org/obo/MONDO_0003927" "http://purl.obolibrary.org/obo/MONDO_0004365" ...
#>   ..$ synonym             :List of 20
#>   ..$ _version_           : num [1:20] 1.6e+18 1.6e+18 1.6e+18 1.6e+18 1.6e+18 ...
#>   ..$ equivalent_iri_kw   :List of 20
#>   ..$ id_kw               : chr [1:20] "MONDO:0006486" "MONDO:0003910" "MONDO:0003927" "MONDO:0004365" ...
#>   ..$ equivalent_curie    :List of 20
#>   ..$ leaf                : logi [1:20] FALSE FALSE TRUE FALSE FALSE TRUE ...
#>   ..$ equivalent_curie_std:List of 20
#>   ..$ equivalent_iri_eng  :List of 20
#>   ..$ iri                 : chr [1:20] "http://purl.obolibrary.org/obo/MONDO_0006486" "http://purl.obolibrary.org/obo/MONDO_0003910" "http://purl.obolibrary.org/obo/MONDO_0003927" "http://purl.obolibrary.org/obo/MONDO_0004365" ...
#>