Expand a graph by fetching connected edges from the KG.

Expand a graph by fetching connected edges from the KG.

fetch_edges(
  g,
  direction = "both",
  predicates = NULL,
  result_categories = NULL,
  transitive = FALSE,
  drop_unused_query_nodes = FALSE
)

Arguments

g

A graph

direction

The direction of associations to fetch. Can be "in", "out", or "both". Default is "both".

predicates

A vector of relationship predicates (nodes in g are subjects in the KG), indicating which edges to consider in the neighborhood. If NULL (default), all edges are considered.

result_categories

A vector of node categories, indicating which nodes in the larger KG (objects in the relationship) to consider as potential result nodes. If NULL (default), all object nodes in the larger KG are considered as potential nodes.

transitive

NOT IMPLEMENTED If TRUE, include transitive closure of the neighborhood. Default is FALSE. Useful in combination with predicates like biolink:subclass_of.

drop_unused_query_nodes

If TRUE, remove query nodes from the result, unless they are at the neighboorhood boundary, ie, required for connecting to the result nodes. Default is FALSE.

Value

A tbl_kgx graph

Examples

g <- monarch_search("fanconi anemia", limit = 1)
phenos <- g |>
  fetch_edges(predicates = "biolink:has_phenotype", result_categories = "biolink:PhenotypicFeature")
#> Error in fetch_edges(g, predicates = "biolink:has_phenotype", result_categories = "biolink:PhenotypicFeature"): could not find function "fetch_edges"

ancestors <- g |>
 fetch_edges(predicates = "biolink:subclass_of", transitive = TRUE, direction = "out")
#> Error in fetch_edges(g, predicates = "biolink:subclass_of", transitive = TRUE,     direction = "out"): could not find function "fetch_edges"