Fetch nodes from a graph using a set of IDs or conditions

This function fetches nodes (and no edges) from a knowledge graph engine based on a set of conditions or a set of identifiers. If query_ids is provided, the function will fetch nodes with the specified identifiers. If query_ids is NULL, the function will fetch nodes based on the conditions provided. Only a limited set of condition expressions are supported, see details.

fetch_nodes(engine, ..., query_ids = NULL, limit = NULL)

Arguments

engine

A graph engine object

...

A set of conditions identifying the nodes to fetch, only used if query_ids is NULL

query_ids

A character vector of identifiers to fetch

limit

An integer specifying the maximum number of nodes to fetch. Default to NULL, no limit.

Value

A tbl_kgx object containing the nodes

Details

If query_ids is provided, the function will fetch nodes with the specified. If query_ids is NULL, the function will fetch nodes based on a condition expression. The following features are supported:

Matching node properties with boolean operators, e.g. in_taxon_label == "Homo sapiens".

Matching multi-valued properties with %in_list%, e.g. "biolink:Gene" %in_list% category. NOTE: using %in_list% against vector queries, e.g. in_taxon_label %in_list% c("Homo sapiens", "Mus musculus") is not supported. Nor does %in_list% support multi-valued left hand sides; c("biolink:Disease", "biolink:Gene") %in_list% category will not work.

Boolean connectives with |, &, and !, e.g. in_taxon_lable == "Homo sapiens" | "biolink:Gene" %in_list% category.

If more than one condition parameter is specified, they are combined with &; for example, fetch_nodes(engine, in_taxon_lable == "Homo sapiens", "biolink:Gene" %in_list% category) is equivalent to fetch_nodes(engine, in_taxon_lable == "Homo sapiens" & "biolink:Gene" %in_list% category).

Examples

library(tidygraph)
#> 
#> Attaching package: ‘tidygraph’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

monarch_engine() |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))
#> Fetching; counting matching nodes... 
#>  total: 2.
#> Fetching; fetched 2 of 2
#> # A tbl_graph: 2 nodes and 0 edges
#> #
#> # A rooted forest with 2 trees
#> #
#> # Node Data: 2 × 10 (active)
#>   id          category pcategory name  description synonym iri   xref  namespace
#>   <chr>       <list>   <chr>     <chr> <chr>       <named> <chr> <nam> <chr>    
#> 1 MONDO:0007… <chr>    biolink:… Ehle… Arthrochal… <chr>   http… <chr> MONDO    
#> 2 MONDO:0007… <chr>    biolink:… Ehle… A form of … <chr>   http… <chr> MONDO    
#> # ℹ 1 more variable: provided_by <named list>
#> #
#> # Edge Data: 0 × 5
#> # ℹ 5 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>

# a large query
monarch_engine() |>
  fetch_nodes("biolink:Disease" %in_list% category)
#> Fetching; counting matching nodes... 
#>  total: 28109.
#> Fetching; fetched 5000 of 28109
#> Fetching; fetched 10000 of 28109
#> Fetching; fetched 15000 of 28109
#> Fetching; fetched 20000 of 28109
#> Fetching; fetched 25000 of 28109
#> Fetching; fetched 28109 of 28109
#> # A tbl_graph: 28109 nodes and 0 edges
#> #
#> # A rooted forest with 28109 trees
#> #
#> # Node Data: 28,109 × 11 (active)
#>    id            category  pcategory   name  description synonym iri   namespace
#>    <chr>         <list>    <chr>       <chr> <chr>       <named> <chr> <chr>    
#>  1 DOID:4        <chr [6]> biolink:Di… NA    NA          <lgl>   http… DOID     
#>  2 MONDO:0000001 <chr [6]> biolink:Di… dise… A disease … <chr>   http… MONDO    
#>  3 MONDO:0000002 <chr [6]> biolink:Di… obso… NA          <lgl>   http… MONDO    
#>  4 MONDO:0000003 <chr [6]> biolink:Di… obso… NA          <lgl>   http… MONDO    
#>  5 MONDO:0000004 <chr [6]> biolink:Di… adre… An endocri… <chr>   http… MONDO    
#>  6 MONDO:0000005 <chr [6]> biolink:Di… alop… NA          <lgl>   http… MONDO    
#>  7 MONDO:0000006 <chr [6]> biolink:Di… obso… NA          <lgl>   http… MONDO    
#>  8 MONDO:0000007 <chr [6]> biolink:Di… obso… NA          <lgl>   http… MONDO    
#>  9 MONDO:0000008 <chr [6]> biolink:Di… obso… NA          <lgl>   http… MONDO    
#> 10 MONDO:0000009 <chr [6]> biolink:Di… inhe… NA          <chr>   http… MONDO    
#> # ℹ 28,099 more rows
#> # ℹ 3 more variables: provided_by <named list>, xref <named list>,
#> #   deprecated <chr>
#> #
#> # Edge Data: 0 × 5
#> # ℹ 5 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>
# file_engine supports the same features as neo4j_engine
# (using the MONDO KGX file packaged with monarchr)
filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")

file_engine(filename) |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))
#> # A tbl_graph: 2 nodes and 0 edges
#> #
#> # A rooted forest with 2 trees
#> #
#> # Node Data: 2 × 11 (active)
#>   id        pcategory name  description synonym category xref  provided_by iri  
#>   <chr>     <chr>     <chr> <chr>       <list>  <list>   <lis> <chr>       <chr>
#> 1 MONDO:00… biolink:… Ehle… Arthrochal… <chr>   <chr>    <chr> mondo.json  http…
#> 2 MONDO:00… biolink:… Ehle… A form of … <chr>   <chr>    <chr> mondo.json  http…
#> # ℹ 2 more variables: same_as <list>, subsets <list>
#> #
#> # Edge Data: 0 × 8
#> # ℹ 8 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>, id <chr>, relation <chr>, knowledge_source <chr>

# grab all rare diseases
file_engine(filename) |>
  fetch_nodes("rare" %in_list% subsets & "biolink:Disease" %in_list% category)
#> # A tbl_graph: 15562 nodes and 0 edges
#> #
#> # A rooted forest with 15562 trees
#> #
#> # Node Data: 15,562 × 11 (active)
#>    id       pcategory name  description synonym category xref  provided_by iri  
#>    <chr>    <chr>     <chr> <chr>       <list>  <list>   <lis> <chr>       <chr>
#>  1 MONDO:0… biolink:… alop… NA          <chr>   <chr>    <chr> mondo.json  http…
#>  2 MONDO:0… biolink:… inhe… NA          <chr>   <chr>    <chr> mondo.json  http…
#>  3 MONDO:0… biolink:… colo… NA          <chr>   <chr>    <chr> mondo.json  http…
#>  4 MONDO:0… biolink:… clas… A genetic … <chr>   <chr>    <chr> mondo.json  http…
#>  5 MONDO:0… biolink:… infa… NA          <chr>   <chr>    <chr> mondo.json  http…
#>  6 MONDO:0… biolink:… slee… NA          <chr>   <chr>    <chr> mondo.json  http…
#>  7 MONDO:0… biolink:… here… Hypophosph… <chr>   <chr>    <chr> mondo.json  http…
#>  8 MONDO:0… biolink:… hypo… NA          <chr>   <chr>    <chr> mondo.json  http…
#>  9 MONDO:0… biolink:… isol… NA          <chr>   <chr>    <chr> mondo.json  http…
#> 10 MONDO:0… biolink:… micr… NA          <chr>   <chr>    <chr> mondo.json  http…
#> # ℹ 15,552 more rows
#> # ℹ 2 more variables: same_as <list>, subsets <list>
#> #
#> # Edge Data: 0 × 8
#> # ℹ 8 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>, id <chr>, relation <chr>, knowledge_source <chr>