Join two KGX graphs by their nodes and edges.

Given two KGX graphs, returns a new KGX graph that is the union of the two input graphs, with any edges between nodes repeated for aother nodes with the same subject and object id. The engine of the first graph is used for the new graph.

kg_join(graph1, graph2, ...)

Arguments

graph1

A tbl_kgx() graph.

graph2

A tbl_kgx() graph.

...

Other parameters (not used)

Value

A tbl_kgx() graph

Details

This function first computes new node and edge data, by taking the full natural join of node and edge data from the two input graphs, and then keeping unique rows. Note that nodes with the same id that differ in any shared column are effectively kept as separate, taken to represent the same entity in different contexts. (However, a node with an additional property will be merged with a node without that property, as defined by the natural join.) In these cases, any edge that connects to one of these nodes is also valid for the other node, and so the method repeats edges across nodes with the same id.

Examples

## Using local MONDO KGX file (packaged with monarchr)
monarch <- file_engine(system.file("extdata", "mondo_kgx_tsv.tar.gz",
                              package = "monarchr"))

eds_and_phenos <- monarch |>
                  fetch_nodes(query_ids = "MONDO:0007525") |>
                  expand(predicates = "biolink:has_phenotype",
                         categories = "biolink:PhenotypicFeature")

marfan_and_phenos <- monarch |>
                     fetch_nodes(query_ids = "MONDO:0007947") |>
                     expand(predicates = "biolink:has_phenotype",
                            categories = "biolink:PhenotypicFeature")

combined <- graph_join(eds_and_phenos, marfan_and_phenos)
#> Joining with `by = join_by(id, pcategory, name, description, synonym, category,
#> xref, provided_by, iri, same_as, subsets)`
print(combined)
#> # A tbl_graph: 2 nodes and 0 edges
#> #
#> # A rooted forest with 2 trees
#> #
#> # Node Data: 2 × 11 (active)
#>   id        pcategory name  description synonym category xref  provided_by iri  
#>   <chr>     <chr>     <chr> <chr>       <list>  <list>   <lis> <chr>       <chr>
#> 1 MONDO:00… biolink:… Ehle… Arthrochal… <chr>   <chr>    <chr> mondo.json  http…
#> 2 MONDO:00… biolink:… Marf… A disorder… <chr>   <chr>    <chr> mondo.json  http…
#> # ℹ 2 more variables: same_as <list>, subsets <list>
#> #
#> # Edge Data: 0 × 5
#> # ℹ 5 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>