Create a knowledge graph engine object for a neo4j database

Creates a knowledge graph engine backed by a neo4j database, from a URL and optional username and password. Knowledge graph "engines" are objects that store information about how to connect to a (potentially large) knowledge graph, and can be used to fetch nodes and edges from the database as local graph objects.

neo4j_engine(url, username = NA, password = NA, preferences = NULL, ...)

Arguments

url

A character string indicating the URL of the neo4j database.

username

A character string indicating the username for the neo4j database (if needed).

password

A character string indicating the password for the neo4j database (if needed).

preferences

A named list of preferences for the engine.

...

Additional arguments passed to neo2R::startGraph().

Value

An object of class neo4j_engine

Details

Engines store preference information specifying how data are fetched and manipulated; for example, while node category is multi-valued (nodes may have multiple categories, for example "biolink:Gene" and "biolink:NamedThing"), typically a single category is used to represent the node in a graph, and is returned as the nodes' pcategory. A preference list of categories to use for pcategory is stored in the engine's preferences. A default set of preferences is stored in the package for use with KGX (BioLink-compatible) graphs (see https://github.com/biolink/kgx/blob/master/specification/kgx-format.md), but these can be overridden by the user.

For neo4j_engine()s, preferences are also used to set the node properties to search when using search_nodes(), defaulting to regex-based searches on id, name, and description. (The monarch_engine() is a type of neo4j_engine() with the URL set to the Monarch Neo4j instance, and overrides search_nodes() to use the Monarch search API, see monarch_engine() for details).

See also

Examples

library(tidygraph)
library(dplyr)

e <- neo4j_engine(url = "https://neo4j.monarchinitiative.org")
print(e$preferences)   # print the default preferences
#> $category_priority
#>  [1] "biolink:Disease"           "biolink:PhenotypicFeature"
#>  [3] "biolink:Gene"              "biolink:AnatomicalEntity" 
#>  [5] "biolink:BiologicalProcess" "biolink:Pathway"          
#>  [7] "biolink:Cell"              "biolink:CellularComponent"
#>  [9] "biolink:ChemicalEntity"    "biolink:Drug"             
#> 
#> $node_property_priority
#> [1] "id"             "pcategory"      "name"           "symbol"        
#> [5] "in_taxon_label" "description"    "synonym"       
#> 
#> $edge_property_priority
#> [1] "subject"   "predicate" "object"   
#> 

# same search and fetch, different preferences
search_nodes(e, "fibrosis", limit = 5) |>
  activate(nodes) |>
  as.data.frame() |>
  select(name, id, pcategory, category)
#> Error in search_nodes(e, "fibrosis", limit = 5): could not find function "search_nodes"

# prefer to set pcategory to "biolink:ThingWithTaxon" if it applies,
# followed by "biolink:NamedThing", otherwise use the first listed category.
# Additionally, only search nodes' name property.
e <- neo4j_engine(url = "https://neo4j.monarchinitiative.org",
                  preferences = list(category_priority = c("biolink:ThingWithTaxon",
                                                           "biolink:NamedThing"),
                                     node_search_properties = c("name")))

search_nodes(e, "fibrosis", limit = 5) |>
  activate(nodes) |>
  as.data.frame() |>
  select(name, id, pcategory, category)
#> Error in search_nodes(e, "fibrosis", limit = 5): could not find function "search_nodes"