Summarize contents of a Neo4j KG engine

Given a Neo4j based KG engine, provides summary information in the form of node counts, category counts across nodes, relationship type counts, and available properties. General information about the graph is printed to the console, and a list of dataframes with this information is returned invisibly. Also returned are cats, preds, and props entries, containing lists of available categories/predicates/properties for convenient auto-completion in RStudio.

# S3 method for class 'neo4j_engine'
summary(object, ..., quiet = FALSE)

Arguments

object

A neo4j_engine object

...

Other parameters (not used)

quiet

Logical, whether to suppress printing of the summary

Value

A list of dataframes and named lists

Examples

# prints a readable summary and returns a list of dataframes
stats <- monarch_engine() |> summary()
#> Trying to connect to https://neo4j.monarchinitiative.org
#> Connected to https://neo4j.monarchinitiative.org
#> 
#> A Neo4j-backed knowledge graph engine.
#> Gathering statistics, please wait...
#> Total nodes:  1379605 
#> Total edges:  15356321 
#> 
#> Node category counts:
#>                                      category   count
#>                            biolink:NamedThing 1379605
#>                                biolink:Entity 1379605
#>                      biolink:BiologicalEntity 1320722
#>                        biolink:ThingWithTaxon 1320722
#>                         biolink:OntologyClass 1286401
#>            biolink:PhysicalEssenceOrOccurrent 1106079
#>                       biolink:PhysicalEssence 1039830
#>                         biolink:GenomicEntity  942858
#>     biolink:ChemicalEntityOrGeneOrGeneProduct  634560
#>                     biolink:GeneOrGeneProduct  611758
#>            biolink:MacromolecularMachineMixin  611758
#>                                  biolink:Gene  585115
#>                       biolink:SequenceVariant  218546
#>            biolink:DiseaseOrPhenotypicFeature  202281
#>                     biolink:PhenotypicFeature  171728
#>                              biolink:Genotype  139197
#>                      biolink:OrganismalEntity   82691
#>                biolink:SubjectOfInvestigation   82691
#>                      biolink:AnatomicalEntity   74170
#>                             biolink:Occurrent   66249
#>           biolink:BiologicalProcessOrActivity   66249
#>                     biolink:BiologicalProcess   53444
#>  biolink:ChemicalEntityOrProteinOrPolypeptide   49445
#>                               biolink:Disease   30553
#>                               biolink:Protein   26643
#>                      biolink:GeneProductMixin   26643
#>                           biolink:Polypeptide   26643
#>                        biolink:ChemicalEntity   22802
#>             biolink:ChemicalOrDrugOrTreatment   22802
#>                               biolink:Pathway   22627
#>                     biolink:MolecularActivity   12805
#>                                  biolink:Case    8207
#>                    biolink:IndividualOrganism    8207
#>                     biolink:CellularComponent    4574
#>                                  biolink:Cell    3449
#>                         biolink:OrganismTaxon    2021
#>                       biolink:MolecularEntity     529
#>                             biolink:LifeStage     314
#> 
#> Edge type counts:
#>                                           predicate   count
#>                              biolink:interacts_with 2758065
#>                                biolink:expressed_in 2443986
#>                               biolink:has_phenotype 2141321
#>                              biolink:orthologous_to 1729200
#>                                     biolink:enables 1305485
#>                        biolink:actively_involved_in 1133499
#>                                  biolink:related_to 1053168
#>                                  biolink:located_in  774267
#>                                 biolink:subclass_of  632420
#>                             biolink:participates_in  346478
#>                  biolink:acts_upstream_of_or_within  234963
#>                      biolink:is_sequence_variant_of  190074
#>                                biolink:is_active_in  177016
#>                        biolink:has_sequence_variant  159036
#>                                     biolink:part_of  144317
#>                                      biolink:causes   19337
#>       biolink:treats_or_applied_or_studied_to_treat   18999
#>                               biolink:homologous_to   14036
#>                              biolink:contributes_to   10599
#>                                    biolink:model_of   10271
#>                     biolink:has_mode_of_inheritance    8820
#>                                 biolink:has_disease    8183
#>                                    biolink:has_gene    8138
#>              biolink:gene_associated_with_condition    8101
#>     biolink:associated_with_increased_likelihood_of    5008
#>                                     biolink:same_as    4734
#>                            biolink:colocalizes_with    4596
#>                            biolink:acts_upstream_of    4038
#>                 biolink:genetically_associated_with    3796
#>                        biolink:disease_has_location    1018
#>  biolink:acts_upstream_of_or_within_positive_effect     800
#>                       biolink:ameliorates_condition     786
#>            biolink:acts_upstream_of_positive_effect     747
#>            biolink:acts_upstream_of_negative_effect     264
#>  biolink:acts_upstream_of_or_within_negative_effect     261
#>                             biolink:has_participant     256
#>                  biolink:preventative_for_condition     145
#>                                    biolink:disrupts      85
#>                                   biolink:caused_by       5
#>                          biolink:contraindicated_in       3
#> 
#> Available node and edge properties:
#>  [1] "id"                               "category"                        
#>  [3] "name"                             "description"                     
#>  [5] "full_name"                        "in_taxon"                        
#>  [7] "in_taxon_label"                   "symbol"                          
#>  [9] "file_source"                      "provided_by"                     
#> [11] "namespace"                        "type"                            
#> [13] "xref"                             "synonym"                         
#> [15] "iri"                              "subsets"                         
#> [17] "exact_synonym"                    "related_synonym"                 
#> [19] "has_gene"                         "deprecated"                      
#> [21] "narrow_synonym"                   "broad_synonym"                   
#> [23] "has_biological_sex"               "predicate"                       
#> [25] "agent_type"                       "aggregator_knowledge_source"     
#> [27] "knowledge_level"                  "primary_knowledge_source"        
#> [29] "publications"                     "has_evidence"                    
#> [31] "subject"                          "object"                          
#> [33] "original_subject"                 "original_object"                 
#> [35] "negated"                          "species_context_qualifier"       
#> [37] "qualifiers"                       "stage_qualifier"                 
#> [39] "disease_context_qualifier"        "has_count"                       
#> [41] "has_percentage"                   "has_quotient"                    
#> [43] "has_total"                        "frequency_qualifier"             
#> [45] "onset_qualifier"                  "sex_qualifier"                   
#> [47] "original_predicate"               "object_specialization_qualifier" 
#> [49] "object_category"                  "subject_category"                
#> [51] "has_attribute"                    "knowledge_source"                
#> [53] "qualifier"                        "subject_specialization_qualifier"
#> 
#> 
#> For more information about Biolink node (Class) and edge (Association) properties, see https://biolink.github.io/biolink-model/.
print(stats)
#> $node_summary
#>                                        category   count
#> 1                            biolink:NamedThing 1379605
#> 2                                biolink:Entity 1379605
#> 3                      biolink:BiologicalEntity 1320722
#> 4                        biolink:ThingWithTaxon 1320722
#> 5                         biolink:OntologyClass 1286401
#> 6            biolink:PhysicalEssenceOrOccurrent 1106079
#> 7                       biolink:PhysicalEssence 1039830
#> 8                         biolink:GenomicEntity  942858
#> 9     biolink:ChemicalEntityOrGeneOrGeneProduct  634560
#> 10                    biolink:GeneOrGeneProduct  611758
#> 11           biolink:MacromolecularMachineMixin  611758
#> 12                                 biolink:Gene  585115
#> 13                      biolink:SequenceVariant  218546
#> 14           biolink:DiseaseOrPhenotypicFeature  202281
#> 15                    biolink:PhenotypicFeature  171728
#> 16                             biolink:Genotype  139197
#> 17                     biolink:OrganismalEntity   82691
#> 18               biolink:SubjectOfInvestigation   82691
#> 19                     biolink:AnatomicalEntity   74170
#> 20                            biolink:Occurrent   66249
#> 21          biolink:BiologicalProcessOrActivity   66249
#> 22                    biolink:BiologicalProcess   53444
#> 23 biolink:ChemicalEntityOrProteinOrPolypeptide   49445
#> 24                              biolink:Disease   30553
#> 25                              biolink:Protein   26643
#> 26                     biolink:GeneProductMixin   26643
#> 27                          biolink:Polypeptide   26643
#> 28                       biolink:ChemicalEntity   22802
#> 29            biolink:ChemicalOrDrugOrTreatment   22802
#> 30                              biolink:Pathway   22627
#> 31                    biolink:MolecularActivity   12805
#> 32                                 biolink:Case    8207
#> 33                   biolink:IndividualOrganism    8207
#> 34                    biolink:CellularComponent    4574
#> 35                                 biolink:Cell    3449
#> 36                        biolink:OrganismTaxon    2021
#> 37                      biolink:MolecularEntity     529
#> 38                            biolink:LifeStage     314
#> 
#> $edge_summary
#>                                             predicate   count
#> 1                              biolink:interacts_with 2758065
#> 2                                biolink:expressed_in 2443986
#> 3                               biolink:has_phenotype 2141321
#> 4                              biolink:orthologous_to 1729200
#> 5                                     biolink:enables 1305485
#> 6                        biolink:actively_involved_in 1133499
#> 7                                  biolink:related_to 1053168
#> 8                                  biolink:located_in  774267
#> 9                                 biolink:subclass_of  632420
#> 10                            biolink:participates_in  346478
#> 11                 biolink:acts_upstream_of_or_within  234963
#> 12                     biolink:is_sequence_variant_of  190074
#> 13                               biolink:is_active_in  177016
#> 14                       biolink:has_sequence_variant  159036
#> 15                                    biolink:part_of  144317
#> 16                                     biolink:causes   19337
#> 17      biolink:treats_or_applied_or_studied_to_treat   18999
#> 18                              biolink:homologous_to   14036
#> 19                             biolink:contributes_to   10599
#> 20                                   biolink:model_of   10271
#> 21                    biolink:has_mode_of_inheritance    8820
#> 22                                biolink:has_disease    8183
#> 23                                   biolink:has_gene    8138
#> 24             biolink:gene_associated_with_condition    8101
#> 25    biolink:associated_with_increased_likelihood_of    5008
#> 26                                    biolink:same_as    4734
#> 27                           biolink:colocalizes_with    4596
#> 28                           biolink:acts_upstream_of    4038
#> 29                biolink:genetically_associated_with    3796
#> 30                       biolink:disease_has_location    1018
#> 31 biolink:acts_upstream_of_or_within_positive_effect     800
#> 32                      biolink:ameliorates_condition     786
#> 33           biolink:acts_upstream_of_positive_effect     747
#> 34           biolink:acts_upstream_of_negative_effect     264
#> 35 biolink:acts_upstream_of_or_within_negative_effect     261
#> 36                            biolink:has_participant     256
#> 37                 biolink:preventative_for_condition     145
#> 38                                   biolink:disrupts      85
#> 39                                  biolink:caused_by       5
#> 40                         biolink:contraindicated_in       3
#> 
#> $total_nodes
#> [1] 1379605
#> 
#> $total_edges
#> [1] 15356321
#> 
#> $cats
#> $cats$`biolink:NamedThing`
#> [1] "biolink:NamedThing"
#> 
#> $cats$`biolink:Entity`
#> [1] "biolink:Entity"
#> 
#> $cats$`biolink:BiologicalEntity`
#> [1] "biolink:BiologicalEntity"
#> 
#> $cats$`biolink:ThingWithTaxon`
#> [1] "biolink:ThingWithTaxon"
#> 
#> $cats$`biolink:OntologyClass`
#> [1] "biolink:OntologyClass"
#> 
#> $cats$`biolink:PhysicalEssenceOrOccurrent`
#> [1] "biolink:PhysicalEssenceOrOccurrent"
#> 
#> $cats$`biolink:PhysicalEssence`
#> [1] "biolink:PhysicalEssence"
#> 
#> $cats$`biolink:GenomicEntity`
#> [1] "biolink:GenomicEntity"
#> 
#> $cats$`biolink:ChemicalEntityOrGeneOrGeneProduct`
#> [1] "biolink:ChemicalEntityOrGeneOrGeneProduct"
#> 
#> $cats$`biolink:GeneOrGeneProduct`
#> [1] "biolink:GeneOrGeneProduct"
#> 
#> $cats$`biolink:MacromolecularMachineMixin`
#> [1] "biolink:MacromolecularMachineMixin"
#> 
#> $cats$`biolink:Gene`
#> [1] "biolink:Gene"
#> 
#> $cats$`biolink:SequenceVariant`
#> [1] "biolink:SequenceVariant"
#> 
#> $cats$`biolink:DiseaseOrPhenotypicFeature`
#> [1] "biolink:DiseaseOrPhenotypicFeature"
#> 
#> $cats$`biolink:PhenotypicFeature`
#> [1] "biolink:PhenotypicFeature"
#> 
#> $cats$`biolink:Genotype`
#> [1] "biolink:Genotype"
#> 
#> $cats$`biolink:OrganismalEntity`
#> [1] "biolink:OrganismalEntity"
#> 
#> $cats$`biolink:SubjectOfInvestigation`
#> [1] "biolink:SubjectOfInvestigation"
#> 
#> $cats$`biolink:AnatomicalEntity`
#> [1] "biolink:AnatomicalEntity"
#> 
#> $cats$`biolink:Occurrent`
#> [1] "biolink:Occurrent"
#> 
#> $cats$`biolink:BiologicalProcessOrActivity`
#> [1] "biolink:BiologicalProcessOrActivity"
#> 
#> $cats$`biolink:BiologicalProcess`
#> [1] "biolink:BiologicalProcess"
#> 
#> $cats$`biolink:ChemicalEntityOrProteinOrPolypeptide`
#> [1] "biolink:ChemicalEntityOrProteinOrPolypeptide"
#> 
#> $cats$`biolink:Disease`
#> [1] "biolink:Disease"
#> 
#> $cats$`biolink:Protein`
#> [1] "biolink:Protein"
#> 
#> $cats$`biolink:GeneProductMixin`
#> [1] "biolink:GeneProductMixin"
#> 
#> $cats$`biolink:Polypeptide`
#> [1] "biolink:Polypeptide"
#> 
#> $cats$`biolink:ChemicalEntity`
#> [1] "biolink:ChemicalEntity"
#> 
#> $cats$`biolink:ChemicalOrDrugOrTreatment`
#> [1] "biolink:ChemicalOrDrugOrTreatment"
#> 
#> $cats$`biolink:Pathway`
#> [1] "biolink:Pathway"
#> 
#> $cats$`biolink:MolecularActivity`
#> [1] "biolink:MolecularActivity"
#> 
#> $cats$`biolink:Case`
#> [1] "biolink:Case"
#> 
#> $cats$`biolink:IndividualOrganism`
#> [1] "biolink:IndividualOrganism"
#> 
#> $cats$`biolink:CellularComponent`
#> [1] "biolink:CellularComponent"
#> 
#> $cats$`biolink:Cell`
#> [1] "biolink:Cell"
#> 
#> $cats$`biolink:OrganismTaxon`
#> [1] "biolink:OrganismTaxon"
#> 
#> $cats$`biolink:MolecularEntity`
#> [1] "biolink:MolecularEntity"
#> 
#> $cats$`biolink:LifeStage`
#> [1] "biolink:LifeStage"
#> 
#> 
#> $preds
#> $preds$`biolink:interacts_with`
#> [1] "biolink:interacts_with"
#> 
#> $preds$`biolink:expressed_in`
#> [1] "biolink:expressed_in"
#> 
#> $preds$`biolink:has_phenotype`
#> [1] "biolink:has_phenotype"
#> 
#> $preds$`biolink:orthologous_to`
#> [1] "biolink:orthologous_to"
#> 
#> $preds$`biolink:enables`
#> [1] "biolink:enables"
#> 
#> $preds$`biolink:actively_involved_in`
#> [1] "biolink:actively_involved_in"
#> 
#> $preds$`biolink:related_to`
#> [1] "biolink:related_to"
#> 
#> $preds$`biolink:located_in`
#> [1] "biolink:located_in"
#> 
#> $preds$`biolink:subclass_of`
#> [1] "biolink:subclass_of"
#> 
#> $preds$`biolink:participates_in`
#> [1] "biolink:participates_in"
#> 
#> $preds$`biolink:acts_upstream_of_or_within`
#> [1] "biolink:acts_upstream_of_or_within"
#> 
#> $preds$`biolink:is_sequence_variant_of`
#> [1] "biolink:is_sequence_variant_of"
#> 
#> $preds$`biolink:is_active_in`
#> [1] "biolink:is_active_in"
#> 
#> $preds$`biolink:has_sequence_variant`
#> [1] "biolink:has_sequence_variant"
#> 
#> $preds$`biolink:part_of`
#> [1] "biolink:part_of"
#> 
#> $preds$`biolink:causes`
#> [1] "biolink:causes"
#> 
#> $preds$`biolink:treats_or_applied_or_studied_to_treat`
#> [1] "biolink:treats_or_applied_or_studied_to_treat"
#> 
#> $preds$`biolink:homologous_to`
#> [1] "biolink:homologous_to"
#> 
#> $preds$`biolink:contributes_to`
#> [1] "biolink:contributes_to"
#> 
#> $preds$`biolink:model_of`
#> [1] "biolink:model_of"
#> 
#> $preds$`biolink:has_mode_of_inheritance`
#> [1] "biolink:has_mode_of_inheritance"
#> 
#> $preds$`biolink:has_disease`
#> [1] "biolink:has_disease"
#> 
#> $preds$`biolink:has_gene`
#> [1] "biolink:has_gene"
#> 
#> $preds$`biolink:gene_associated_with_condition`
#> [1] "biolink:gene_associated_with_condition"
#> 
#> $preds$`biolink:associated_with_increased_likelihood_of`
#> [1] "biolink:associated_with_increased_likelihood_of"
#> 
#> $preds$`biolink:same_as`
#> [1] "biolink:same_as"
#> 
#> $preds$`biolink:colocalizes_with`
#> [1] "biolink:colocalizes_with"
#> 
#> $preds$`biolink:acts_upstream_of`
#> [1] "biolink:acts_upstream_of"
#> 
#> $preds$`biolink:genetically_associated_with`
#> [1] "biolink:genetically_associated_with"
#> 
#> $preds$`biolink:disease_has_location`
#> [1] "biolink:disease_has_location"
#> 
#> $preds$`biolink:acts_upstream_of_or_within_positive_effect`
#> [1] "biolink:acts_upstream_of_or_within_positive_effect"
#> 
#> $preds$`biolink:ameliorates_condition`
#> [1] "biolink:ameliorates_condition"
#> 
#> $preds$`biolink:acts_upstream_of_positive_effect`
#> [1] "biolink:acts_upstream_of_positive_effect"
#> 
#> $preds$`biolink:acts_upstream_of_negative_effect`
#> [1] "biolink:acts_upstream_of_negative_effect"
#> 
#> $preds$`biolink:acts_upstream_of_or_within_negative_effect`
#> [1] "biolink:acts_upstream_of_or_within_negative_effect"
#> 
#> $preds$`biolink:has_participant`
#> [1] "biolink:has_participant"
#> 
#> $preds$`biolink:preventative_for_condition`
#> [1] "biolink:preventative_for_condition"
#> 
#> $preds$`biolink:disrupts`
#> [1] "biolink:disrupts"
#> 
#> $preds$`biolink:caused_by`
#> [1] "biolink:caused_by"
#> 
#> $preds$`biolink:contraindicated_in`
#> [1] "biolink:contraindicated_in"
#> 
#> 
#> $props
#> $props$id
#> [1] "id"
#> 
#> $props$category
#> [1] "category"
#> 
#> $props$name
#> [1] "name"
#> 
#> $props$description
#> [1] "description"
#> 
#> $props$full_name
#> [1] "full_name"
#> 
#> $props$in_taxon
#> [1] "in_taxon"
#> 
#> $props$in_taxon_label
#> [1] "in_taxon_label"
#> 
#> $props$symbol
#> [1] "symbol"
#> 
#> $props$file_source
#> [1] "file_source"
#> 
#> $props$provided_by
#> [1] "provided_by"
#> 
#> $props$namespace
#> [1] "namespace"
#> 
#> $props$type
#> [1] "type"
#> 
#> $props$xref
#> [1] "xref"
#> 
#> $props$synonym
#> [1] "synonym"
#> 
#> $props$iri
#> [1] "iri"
#> 
#> $props$subsets
#> [1] "subsets"
#> 
#> $props$exact_synonym
#> [1] "exact_synonym"
#> 
#> $props$related_synonym
#> [1] "related_synonym"
#> 
#> $props$has_gene
#> [1] "has_gene"
#> 
#> $props$deprecated
#> [1] "deprecated"
#> 
#> $props$narrow_synonym
#> [1] "narrow_synonym"
#> 
#> $props$broad_synonym
#> [1] "broad_synonym"
#> 
#> $props$has_biological_sex
#> [1] "has_biological_sex"
#> 
#> $props$predicate
#> [1] "predicate"
#> 
#> $props$agent_type
#> [1] "agent_type"
#> 
#> $props$aggregator_knowledge_source
#> [1] "aggregator_knowledge_source"
#> 
#> $props$knowledge_level
#> [1] "knowledge_level"
#> 
#> $props$primary_knowledge_source
#> [1] "primary_knowledge_source"
#> 
#> $props$publications
#> [1] "publications"
#> 
#> $props$has_evidence
#> [1] "has_evidence"
#> 
#> $props$subject
#> [1] "subject"
#> 
#> $props$object
#> [1] "object"
#> 
#> $props$original_subject
#> [1] "original_subject"
#> 
#> $props$original_object
#> [1] "original_object"
#> 
#> $props$negated
#> [1] "negated"
#> 
#> $props$species_context_qualifier
#> [1] "species_context_qualifier"
#> 
#> $props$qualifiers
#> [1] "qualifiers"
#> 
#> $props$stage_qualifier
#> [1] "stage_qualifier"
#> 
#> $props$disease_context_qualifier
#> [1] "disease_context_qualifier"
#> 
#> $props$has_count
#> [1] "has_count"
#> 
#> $props$has_percentage
#> [1] "has_percentage"
#> 
#> $props$has_quotient
#> [1] "has_quotient"
#> 
#> $props$has_total
#> [1] "has_total"
#> 
#> $props$frequency_qualifier
#> [1] "frequency_qualifier"
#> 
#> $props$onset_qualifier
#> [1] "onset_qualifier"
#> 
#> $props$sex_qualifier
#> [1] "sex_qualifier"
#> 
#> $props$original_predicate
#> [1] "original_predicate"
#> 
#> $props$object_specialization_qualifier
#> [1] "object_specialization_qualifier"
#> 
#> $props$object_category
#> [1] "object_category"
#> 
#> $props$subject_category
#> [1] "subject_category"
#> 
#> $props$has_attribute
#> [1] "has_attribute"
#> 
#> $props$knowledge_source
#> [1] "knowledge_source"
#> 
#> $props$qualifier
#> [1] "qualifier"
#> 
#> $props$subject_specialization_qualifier
#> [1] "subject_specialization_qualifier"
#> 
#>