Transfer information over edges to nodes.

Used to 'transfer' information from nodes to other nodes across specific predicates, either in an outward direction (along the edge direction) or inward (against the edge direction). Returns a node-property column; intended to be used with mutate() on nodes.

transfer(colname = NULL, over, direction = "out")

Arguments

colname

The node column to transfer information from

over

The edge predicate to transfer information over

direction

Whether to transfer information along the predicate direction ("out") or against ("in")

Value

Vector or list, with one entry per node.

Details

The return value will be either a list, or if the result would be a list with all length-1 or length-0 elements, a vector with 0-length elements replaced by NA. Practically, this results in a list when necessary and a vector otherwise.

See also

Examples