Return an example set of nodes from a KG engine.

Given a KG engine, returns a graph representing the diversity of node categories and edge predicates for browsing. The returned graph is guaranteed to contain at least one node of every category, and at least one edge of every predicate. No other guarantees are made: the example graph is not minimal to satisfy these criteria, it is not random or even pseudo-random, and it may not be connected.

example_graph(engine, ...)

Arguments

engine

A KG engine object

...

Other parameters (not used)

Value

A tbl_kgx graph

Examples

# Using example KGX file packaged with monarchr
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")

# prints a readable summary and returns a list of dataframes
g <- file_engine(filename) |> example_graph()
print(g)
#> # A tbl_graph: 12 nodes and 10 edges
#> #
#> # A directed acyclic multigraph with 3 components
#> #
#> # Node Data: 12 × 16 (active)
#>    id         pcategory name  symbol in_taxon_label description synonym category
#>    <chr>      <chr>     <chr> <chr>  <chr>          <chr>       <list>  <list>  
#>  1 MONDO:000… biolink:… Marf… NA     NA             A disorder… <chr>   <chr>   
#>  2 MONDO:002… biolink:… Ehle… NA     NA             The Ehlers… <chr>   <chr>   
#>  3 MONDO:000… biolink:… Ehle… NA     NA             Ehlers-Dan… <chr>   <chr>   
#>  4 MONDO:001… biolink:… Ehle… NA     NA             A form of … <chr>   <chr>   
#>  5 ZFIN:ZDB-… biolink:… b3ga… NA     Danio rerio    NA          <chr>   <chr>   
#>  6 HGNC:3603  biolink:… FBN1  FBN1   Homo sapiens   NA          <chr>   <chr>   
#>  7 HP:0000974 biolink:… Hype… NA     NA             A conditio… <chr>   <chr>   
#>  8 HP:0000007 biolink:… Auto… NA     NA             A mode of … <chr>   <chr>   
#>  9 MONDO:002… biolink:… rare  NA     NA             A disease … <chr>   <chr>   
#> 10 CHEBI:508… biolink:… doxy… NA     NA             Tetracycli… <chr>   <chr>   
#> 11 CLINVAR:2… biolink:… NM_0… NA     Homo sapiens   NA          <chr>   <chr>   
#> 12 CLINVAR:2… biolink:… NM_0… NA     Homo sapiens   NA          <chr>   <chr>   
#> # ℹ 8 more variables: iri <chr>, xref <list>, namespace <chr>,
#> #   provided_by <chr>, in_taxon <chr>, full_name <chr>, type <list>,
#> #   has_gene <chr>
#> #
#> # Edge Data: 10 × 25
#>    from    to subject        predicate         object agent_type knowledge_level
#>   <int> <int> <chr>          <chr>             <chr>  <chr>      <chr>          
#> 1    11     1 CLINVAR:200198 biolink:associat… MONDO… manual_ag… knowledge_asse…
#> 2     6     1 HGNC:3603      biolink:causes    MONDO… manual_ag… knowledge_asse…
#> 3     6     1 HGNC:3603      biolink:gene_ass… MONDO… manual_ag… knowledge_asse…
#> # ℹ 7 more rows
#> # ℹ 18 more variables: knowledge_source <chr>,
#> #   aggregator_knowledge_source <chr>, primary_knowledge_source <chr>,
#> #   provided_by <chr>, id <chr>, category <chr>, original_object <chr>,
#> #   original_subject <chr>, frequency_qualifier <chr>, has_evidence <chr>,
#> #   has_total <dbl>, has_quotient <dbl>, has_count <dbl>, has_percentage <dbl>,
#> #   onset_qualifier <chr>, publications <chr>, qualifiers <chr>, …

# prints a readable summary and returns a list of dataframes
g <- monarch_engine() |> example_graph()
#> Trying to connect to http://neo4j.monarchinitiative.org:7474
#> Connected to http://neo4j.monarchinitiative.org:7474
print(g)
#> # A tbl_graph: 89 nodes and 50 edges
#> #
#> # A directed acyclic multigraph with 40 components
#> #
#> # Node Data: 89 × 17 (active)
#>    id         category pcategory name  symbol in_taxon_label description synonym
#>    <chr>      <list>   <chr>     <chr> <chr>  <chr>          <chr>       <named>
#>  1 GO:0045944 <chr>    biolink:… posi… NA     NA             Any proces… <chr>  
#>  2 WB:WBGene… <chr>    biolink:… lin-… lin-41 Caenorhabditi… NA          <chr>  
#>  3 FB:FBgn00… <chr>    biolink:… Rbfo… Rbfox1 Drosophila me… NA          <chr>  
#>  4 WB:WBGene… <chr>    biolink:… spn-4 spn-4  Caenorhabditi… NA          <chr>  
#>  5 GO:1990124 <chr>    biolink:… mess… NA     NA             A ribonucl… <chr>  
#>  6 WB:WBGene… <chr>    biolink:… ceh-… ceh-18 Caenorhabditi… NA          <chr>  
#>  7 GO:0001556 <chr>    biolink:… oocy… NA     NA             A developm… <lgl>  
#>  8 GO:0005634 <chr>    biolink:… nucl… NA     NA             A membrane… <chr>  
#>  9 CLINVAR:1… <chr>    biolink:… NM_0… NA     Homo sapiens   NA          <lgl>  
#> 10 MONDO:010… <chr>    biolink:… over… NA     NA             A disease … <chr>  
#> # ℹ 79 more rows
#> # ℹ 9 more variables: iri <chr>, namespace <chr>, provided_by <named list>,
#> #   full_name <chr>, xref <named list>, in_taxon <named list>,
#> #   type <named list>, has_gene <named list>, deprecated <chr>
#> #
#> # Edge Data: 50 × 21
#>    from    to subject           predicate         object knowledge_level negated
#>   <int> <int> <chr>             <chr>             <chr>  <chr>           <lgl>  
#> 1     2     1 WB:WBGene00003026 biolink:acts_ups… GO:00… knowledge_asse… TRUE   
#> 2     4     3 WB:WBGene00004984 biolink:ortholog… FB:FB… knowledge_asse… NA     
#> 3     2     5 WB:WBGene00003026 biolink:part_of   GO:19… knowledge_asse… TRUE   
#> # ℹ 47 more rows
#> # ℹ 14 more variables: primary_knowledge_source <chr>, agent_type <chr>,
#> #   knowledge_source <chr>, aggregator_knowledge_source <named list>,
#> #   has_evidence <named list>, provided_by <named list>, id <chr>,
#> #   category <named list>, qualifiers <named list>, original_predicate <chr>,
#> #   original_object <chr>, original_subject <chr>, publications <named list>,
#> #   stage_qualifier <chr>