Fetch additional knowledge graph edges connected to a query graph

Given an optional KG engine (e.g. a file_engine(), neo4j_engine(), or monarch_engine()) and a query tbl_kgx() graph, fetches additional nodes and edges from the KG, expanding the query graph according to specific criteria. If the first parameter is an engine, that engine is used; if the first parameter is a query graph, the most recent engine associated with the graph is used.

expand(
  graph,
  engine = NULL,
  direction = "both",
  predicates = NULL,
  categories = NULL,
  transitive = FALSE,
  drop_unused_query_nodes = FALSE,
  ...
)

Arguments

graph

A query tbl_kgx() graph ot query with.

engine

(Optional) An engine to use for fetching query graph edges. If not provided, the graph's most recent engine is used.

direction

The direction of associations to fetch. Can be "in", "out", or "both". Default is "both".

predicates

A vector of relationship predicates (nodes in g are subjects in the KG), indicating which edges to consider in the neighborhood. If NULL (default), all edges are considered.

categories

A vector of node categories, indicating which nodes in the larger KG may be fetched. If NULL (default), all nodes in the larger KG are will be fetched.

transitive

If TRUE, include transitive closure of the neighborhood. Default is FALSE. Useful in combination with predicates like biolink:subclass_of.

drop_unused_query_nodes

If TRUE, remove query nodes from the result, unless they are at the neighborhood boundary, i.e., required for connecting to the result nodes. Default is FALSE.

...

Other parameters passed to methods.

Value

A tbl_kgx() graph

Examples

## Using Monarch (hosted)
phenos <- monarch_engine() |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")
#> Fetching; counting matching nodes... 
#>  total: 1.
#> Fetching; fetched 1 of 1
#> Expanding; counting matching edges... 
#>  total: 48.
#> Expanding; fetched 48 of 48 edges.

print(phenos)
#> # A tbl_graph: 49 nodes and 48 edges
#> #
#> # A rooted tree
#> #
#> # Node Data: 49 × 10 (active)
#>    id         category pcategory name  description synonym iri   xref  namespace
#>    <chr>      <list>   <chr>     <chr> <chr>       <named> <chr> <nam> <chr>    
#>  1 MONDO:000… <chr>    biolink:… Ehle… Arthrochal… <chr>   http… <chr> MONDO    
#>  2 HP:0001058 <chr>    biolink:… Poor… A reduced … <chr>   http… <chr> HP       
#>  3 HP:0001476 <chr>    biolink:… Dela… A delay in… <chr>   http… <chr> HP       
#>  4 HP:0001252 <chr>    biolink:… Hypo… Hypotonia … <chr>   http… <chr> HP       
#>  5 HP:0001075 <chr>    biolink:… Atro… Scars that… <chr>   http… <chr> HP       
#>  6 HP:0001290 <chr>    biolink:… Gene… Generalize… <chr>   http… <chr> HP       
#>  7 HP:0005191 <chr>    biolink:… Cong… NA          <chr>   http… <chr> HP       
#>  8 HP:0009540 <chr>    biolink:… Cont… Chronic lo… <chr>   http… <chr> HP       
#>  9 HP:0000938 <chr>    biolink:… Oste… Osteopenia… <chr>   http… <chr> HP       
#> 10 HP:0001822 <chr>    biolink:… Hall… Lateral de… <chr>   http… <chr> HP       
#> # ℹ 39 more rows
#> # ℹ 1 more variable: provided_by <named list>
#> #
#> # Edge Data: 48 × 22
#>    from    to subject    predicate object knowledge_level primary_knowledge_so…¹
#>   <int> <int> <chr>      <chr>     <chr>  <chr>           <chr>                 
#> 1     1     2 MONDO:000… biolink:… HP:00… knowledge_asse… infores:hpo-annotatio…
#> 2     1     3 MONDO:000… biolink:… HP:00… knowledge_asse… infores:hpo-annotatio…
#> 3     1     4 MONDO:000… biolink:… HP:00… knowledge_asse… infores:hpo-annotatio…
#> # ℹ 45 more rows
#> # ℹ abbreviated name: ¹​primary_knowledge_source
#> # ℹ 15 more variables: original_subject <chr>, agent_type <chr>,
#> #   knowledge_source <chr>, aggregator_knowledge_source <named list>,
#> #   has_evidence <named list>, provided_by <named list>, id <chr>,
#> #   category <named list>, has_total <chr>, has_quotient <chr>,
#> #   has_count <chr>, has_percentage <chr>, publications <named list>, …

## Using local MONDO KGX file (packaged with monarchr)
phenos <- file_engine(system.file("extdata", "mondo_kgx_tsv.tar.gz",
                      package = "monarchr")) |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")

print(phenos)
#> # A tbl_graph: 1 nodes and 0 edges
#> #
#> # A rooted tree
#> #
#> # Edge Data: 0 × 8 (active)
#> # ℹ 8 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>, id <chr>, relation <chr>, knowledge_source <chr>
#> #
#> # Node Data: 1 × 11
#>   id        pcategory name  description synonym category xref  provided_by iri  
#>   <chr>     <chr>     <chr> <chr>       <list>  <list>   <lis> <chr>       <chr>
#> 1 MONDO:00… biolink:… Ehle… Arthrochal… <chr>   <chr>    <chr> mondo.json  http…
#> # ℹ 2 more variables: same_as <list>, subsets <list>

## Using MONDO KGX file (remote) as an example
phenos <- file_engine("https://kghub.io/kg-obo/mondo/2024-03-04/mondo_kgx_tsv.tar.gz") |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")

print(phenos)
#> # A tbl_graph: 1 nodes and 0 edges
#> #
#> # A rooted tree
#> #
#> # Edge Data: 0 × 8 (active)
#> # ℹ 8 variables: from <int>, to <int>, subject <chr>, predicate <chr>,
#> #   object <chr>, id <chr>, relation <chr>, knowledge_source <chr>
#> #
#> # Node Data: 1 × 11
#>   id        pcategory name  description synonym category xref  provided_by iri  
#>   <chr>     <chr>     <chr> <chr>       <list>  <list>   <lis> <chr>       <chr>
#> 1 MONDO:00… biolink:… Ehle… Arthrochal… <chr>   <chr>    <chr> mondo.json  http…
#> # ℹ 2 more variables: same_as <list>, subsets <list>

file.remove("mondo_kgx_tsv.tar.gz") # cleanup - remove the downloaded file
#> [1] TRUE