Summarize neighborhood

Summarizes the neighborhood of the nodes of a given graph. Specifically, letting $N$ be the set of nodes in the given graph, this function retrieves counts of relationship predicates (or node categories) of nodes connected to $N$ but not in $N$. This can be useful to examine the scale and scope of a graph's collective neighborhood in the larger KG.

summarize_neighborhood(g, direction = "both", summarize = "edges")

Arguments

g

A graph

direction

The direction of edges to include in the neighborhood

summarize

Whether to summarize edges, nodes, or both

Value

A tbl_kgx graph

Details

Note that the number of relationships returned may be larger than the number of nodes they connect to; use summarize = "edges" to see edge counts between nodes of different categories, and summarize = "nodes" to see counts of connected node categories.

It is also possible to specify the direction of edges to include in the neighborhood, using the direction parameter. The default is "both", which includes both incoming and outgoing edges.

Examples

g <- monarch_search("fanconi anemia")
summarize_neighborhood(g, summarize = "edges") # default
#> # A tibble: 7 × 6
#>   count query_pcategory           predicate      result_pcategory query_category
#>   <int> <chr>                     <chr>          <chr>            <list>        
#> 1   126 biolink:BiologicalProcess biolink:parti… biolink:Gene     <list [10]>   
#> 2   106 biolink:Disease           biolink:has_p… biolink:Phenoty… <list [6]>    
#> 3    30 biolink:Disease           biolink:subcl… biolink:Disease  <list [6]>    
#> 4    23 biolink:BiologicalProcess biolink:parti… biolink:Chemica… <list [10]>   
#> 5    22 biolink:Disease           biolink:gene_… biolink:Gene     <list [6]>    
#> 6     2 biolink:Disease           biolink:treat… biolink:Chemica… <list [6]>    
#> 7     1 biolink:Disease           biolink:relat… biolink:Disease  <list [6]>    
#> # ℹ 1 more variable: result_category <list>
summarize_neighborhood(g, summarize = "nodes")
#> # A tibble: 4 × 2
#>   pcategory                 count
#>   <chr>                     <int>
#> 1 biolink:Gene                148
#> 2 biolink:PhenotypicFeature   106
#> 3 biolink:Disease              31
#> 4 biolink:ChemicalEntity       25