Summarize contents of a Neo4j KG engine

Given a Neo4j based KG engine, provides summary information in the form of node counts, category counts across nodes, and relationship type counts. General information about the graph is printed to the console, and a list of dataframes describing node and edge counts is returned invisibly. Also returned are cats and preds entries, containing lists of available node categories and edge predicates, respectively, for convenient auto-completion in RStudio.

# S3 method for class 'neo4j_engine'
summary(object, ..., quiet = FALSE)

Arguments

object

A neo4j_engine object

...

Other parameters (not used)

quiet

Logical, whether to suppress printing of the summary

Value

A list of dataframes and named lists

Examples

# prints a readable summary and returns a list of dataframes
stats <- monarch_engine() |> summary()
#> Trying to connect to http://neo4j.monarchinitiative.org:7474
#> Connected to http://neo4j.monarchinitiative.org:7474
#> 
#> A Neo4j-backed knowledge graph engine.
#> Gathering statistics, please wait...
#> Total nodes:  1067161 
#> Total edges:  8627308 
#> 
#> Node category counts:
#>                                      category   count
#>                                biolink:Entity 1067161
#>                            biolink:NamedThing 1067161
#>                        biolink:ThingWithTaxon 1029960
#>                      biolink:BiologicalEntity 1029942
#>            biolink:PhysicalEssenceOrOccurrent  855402
#>                       biolink:PhysicalEssence  789389
#>                         biolink:OntologyClass  785207
#>                         biolink:GenomicEntity  719054
#>     biolink:ChemicalEntityOrGeneOrGeneProduct  576838
#>            biolink:MacromolecularMachineMixin  574382
#>                     biolink:GeneOrGeneProduct  572267
#>                                  biolink:Gene  571134
#>            biolink:DiseaseOrPhenotypicFeature  173742
#>                     biolink:PhenotypicFeature  144269
#>                              biolink:Genotype  134888
#>                biolink:SubjectOfInvestigation   67895
#>                      biolink:OrganismalEntity   67894
#>                             biolink:Occurrent   66013
#>           biolink:BiologicalProcessOrActivity   66009
#>                      biolink:AnatomicalEntity   65773
#>                               biolink:Disease   29473
#>                     biolink:BiologicalProcess   26268
#>                                  biolink:Cell   26176
#>              biolink:GrossAnatomicalStructure   24329
#>                               biolink:Pathway   22331
#>                       biolink:SequenceVariant   13007
#>  biolink:ChemicalEntityOrProteinOrPolypeptide    5698
#>                     biolink:CellularComponent    5319
#>             biolink:ChemicalOrDrugOrTreatment    4580
#>                        biolink:ChemicalEntity    4579
#>                       biolink:MolecularEntity    4242
#>                 biolink:MacromolecularComplex    2115
#>                     biolink:MolecularActivity    1560
#>                      biolink:CellularOrganism    1555
#>                      biolink:GeneProductMixin    1133
#>                           biolink:Polypeptide    1119
#>                               biolink:Protein    1118
#>                            biolink:Vertebrate     584
#>                                 biolink:Virus     322
#>                     biolink:BehavioralFeature     297
#>                             biolink:LifeStage     238
#>                   biolink:PathologicalProcess     231
#>               biolink:PathologicalEntityMixin     231
#>                       biolink:ChemicalMixture      97
#>                      biolink:MolecularMixture      93
#>                                  biolink:Drug      93
#>                         biolink:SmallMolecule      60
#>                             biolink:Attribute      50
#>              biolink:InformationContentEntity      49
#>                     biolink:ClinicalAttribute      44
#>                                 biolink:Onset      44
#>                        biolink:ClinicalCourse      44
#>                         biolink:OrganismTaxon      33
#>                            biolink:Transcript      21
#>                     biolink:NucleicAcidEntity      18
#>                          biolink:EvidenceType      16
#>                            biolink:RNAProduct      15
#>                                 biolink:Plant       4
#>                           biolink:Publication       4
#>                                biolink:Fungus       4
#>                         biolink:ExposureEvent       3
#>                       biolink:PlanetaryEntity       3
#>                     biolink:ProcessedMaterial       3
#>                              biolink:Activity       3
#>                   biolink:ActivityAndBehavior       3
#>                                biolink:Mammal       3
#>                     biolink:GeneGroupingMixin       3
#>                         biolink:BiologicalSex       3
#>                      biolink:RegulatoryRegion       3
#>                      biolink:ChemicalExposure       2
#>                  biolink:EnvironmentalFeature       2
#>                  biolink:AdministrativeEntity       2
#>                                 biolink:Agent       2
#>                               biolink:Dataset       2
#>                       biolink:ConfidenceLevel       2
#>                          biolink:Invertebrate       2
#>                    biolink:GeneticInheritance       2
#>                             biolink:Haplotype       2
#>                         biolink:ProteinFamily       2
#>       biolink:PopulationOfIndividualOrganisms       2
#>                   biolink:NoncodingRNAProduct       2
#>                          biolink:DrugExposure       1
#>                               biolink:WebPage       1
#>                   biolink:DatasetDistribution       1
#>                    biolink:IndividualOrganism       1
#>                                biolink:Patent       1
#>                              biolink:CellLine       1
#>                  biolink:EnvironmentalProcess       1
#>                              biolink:Zygosity       1
#>                   biolink:ReagentTargetedGene       1
#>                             biolink:Bacterium       1
#>                         biolink:ProteinDomain       1
#>                         biolink:StudyVariable       1
#>                                 biolink:Human       1
#>                                 biolink:Event       1
#>                                 biolink:Study       1
#>                        biolink:MaterialSample       1
#>                        biolink:PhysicalEntity       1
#>                             biolink:Treatment       1
#>                         biolink:PhenotypicSex       1
#>                          biolink:GenotypicSex       1
#>                                  biolink:Exon       1
#>        biolink:TranscriptionFactorBindingSite       1
#>                              biolink:MicroRNA       1
#>                                 biolink:SiRNA       1
#>                                biolink:Genome       1
#>                                   biolink:Snv       1
#>                   biolink:AccessibleDnaRegion       1
#>                         biolink:DiagnosticAid       1
#> 
#> Edge type counts:
#>                                           predicate   count
#>                              biolink:interacts_with 2483431
#>                               biolink:has_phenotype 1482455
#>                                biolink:expressed_in 1229165
#>                        biolink:actively_involved_in  632984
#>                              biolink:orthologous_to  551358
#>                                 biolink:subclass_of  544535
#>                                     biolink:enables  431854
#>                                  biolink:located_in  312598
#>                             biolink:participates_in  268142
#>                                  biolink:related_to  256225
#>                  biolink:acts_upstream_of_or_within  145303
#>                                   biolink:active_in  144810
#>                                     biolink:part_of   63174
#>                                      biolink:causes   15651
#>                      biolink:is_sequence_variant_of   13100
#>                                    biolink:model_of    9258
#>                     biolink:has_mode_of_inheritance    8205
#>              biolink:gene_associated_with_condition    8052
#>                            biolink:acts_upstream_of    7095
#>                              biolink:contributes_to    6332
#>       biolink:treats_or_applied_or_studied_to_treat    5335
#>                            biolink:colocalizes_with    2653
#>     biolink:associated_with_increased_likelihood_of    2208
#>                 biolink:genetically_associated_with    2155
#>            biolink:acts_upstream_of_positive_effect     473
#>  biolink:acts_upstream_of_or_within_positive_effect     440
#>            biolink:acts_upstream_of_negative_effect     165
#>  biolink:acts_upstream_of_or_within_negative_effect     152
print(stats)
#> $node_summary
#>                                         category   count
#> 1                                 biolink:Entity 1067161
#> 2                             biolink:NamedThing 1067161
#> 3                         biolink:ThingWithTaxon 1029960
#> 4                       biolink:BiologicalEntity 1029942
#> 5             biolink:PhysicalEssenceOrOccurrent  855402
#> 6                        biolink:PhysicalEssence  789389
#> 7                          biolink:OntologyClass  785207
#> 8                          biolink:GenomicEntity  719054
#> 9      biolink:ChemicalEntityOrGeneOrGeneProduct  576838
#> 10            biolink:MacromolecularMachineMixin  574382
#> 11                     biolink:GeneOrGeneProduct  572267
#> 12                                  biolink:Gene  571134
#> 13            biolink:DiseaseOrPhenotypicFeature  173742
#> 14                     biolink:PhenotypicFeature  144269
#> 15                              biolink:Genotype  134888
#> 16                biolink:SubjectOfInvestigation   67895
#> 17                      biolink:OrganismalEntity   67894
#> 18                             biolink:Occurrent   66013
#> 19           biolink:BiologicalProcessOrActivity   66009
#> 20                      biolink:AnatomicalEntity   65773
#> 21                               biolink:Disease   29473
#> 22                     biolink:BiologicalProcess   26268
#> 23                                  biolink:Cell   26176
#> 24              biolink:GrossAnatomicalStructure   24329
#> 25                               biolink:Pathway   22331
#> 26                       biolink:SequenceVariant   13007
#> 27  biolink:ChemicalEntityOrProteinOrPolypeptide    5698
#> 28                     biolink:CellularComponent    5319
#> 29             biolink:ChemicalOrDrugOrTreatment    4580
#> 30                        biolink:ChemicalEntity    4579
#> 31                       biolink:MolecularEntity    4242
#> 32                 biolink:MacromolecularComplex    2115
#> 33                     biolink:MolecularActivity    1560
#> 34                      biolink:CellularOrganism    1555
#> 35                      biolink:GeneProductMixin    1133
#> 36                           biolink:Polypeptide    1119
#> 37                               biolink:Protein    1118
#> 38                            biolink:Vertebrate     584
#> 39                                 biolink:Virus     322
#> 40                     biolink:BehavioralFeature     297
#> 41                             biolink:LifeStage     238
#> 42                   biolink:PathologicalProcess     231
#> 43               biolink:PathologicalEntityMixin     231
#> 44                       biolink:ChemicalMixture      97
#> 45                      biolink:MolecularMixture      93
#> 46                                  biolink:Drug      93
#> 47                         biolink:SmallMolecule      60
#> 48                             biolink:Attribute      50
#> 49              biolink:InformationContentEntity      49
#> 50                     biolink:ClinicalAttribute      44
#> 51                                 biolink:Onset      44
#> 52                        biolink:ClinicalCourse      44
#> 53                         biolink:OrganismTaxon      33
#> 54                            biolink:Transcript      21
#> 55                     biolink:NucleicAcidEntity      18
#> 56                          biolink:EvidenceType      16
#> 57                            biolink:RNAProduct      15
#> 58                                 biolink:Plant       4
#> 59                           biolink:Publication       4
#> 60                                biolink:Fungus       4
#> 61                         biolink:ExposureEvent       3
#> 62                       biolink:PlanetaryEntity       3
#> 63                     biolink:ProcessedMaterial       3
#> 64                              biolink:Activity       3
#> 65                   biolink:ActivityAndBehavior       3
#> 66                                biolink:Mammal       3
#> 67                     biolink:GeneGroupingMixin       3
#> 68                         biolink:BiologicalSex       3
#> 69                      biolink:RegulatoryRegion       3
#> 70                      biolink:ChemicalExposure       2
#> 71                  biolink:EnvironmentalFeature       2
#> 72                  biolink:AdministrativeEntity       2
#> 73                                 biolink:Agent       2
#> 74                               biolink:Dataset       2
#> 75                       biolink:ConfidenceLevel       2
#> 76                          biolink:Invertebrate       2
#> 77                    biolink:GeneticInheritance       2
#> 78                             biolink:Haplotype       2
#> 79                         biolink:ProteinFamily       2
#> 80       biolink:PopulationOfIndividualOrganisms       2
#> 81                   biolink:NoncodingRNAProduct       2
#> 82                          biolink:DrugExposure       1
#> 83                               biolink:WebPage       1
#> 84                   biolink:DatasetDistribution       1
#> 85                    biolink:IndividualOrganism       1
#> 86                                biolink:Patent       1
#> 87                              biolink:CellLine       1
#> 88                  biolink:EnvironmentalProcess       1
#> 89                              biolink:Zygosity       1
#> 90                   biolink:ReagentTargetedGene       1
#> 91                             biolink:Bacterium       1
#> 92                         biolink:ProteinDomain       1
#> 93                         biolink:StudyVariable       1
#> 94                                 biolink:Human       1
#> 95                                 biolink:Event       1
#> 96                                 biolink:Study       1
#> 97                        biolink:MaterialSample       1
#> 98                        biolink:PhysicalEntity       1
#> 99                             biolink:Treatment       1
#> 100                        biolink:PhenotypicSex       1
#> 101                         biolink:GenotypicSex       1
#> 102                                 biolink:Exon       1
#> 103       biolink:TranscriptionFactorBindingSite       1
#> 104                             biolink:MicroRNA       1
#> 105                                biolink:SiRNA       1
#> 106                               biolink:Genome       1
#> 107                                  biolink:Snv       1
#> 108                  biolink:AccessibleDnaRegion       1
#> 109                        biolink:DiagnosticAid       1
#> 
#> $edge_summary
#>                                             predicate   count
#> 1                              biolink:interacts_with 2483431
#> 2                               biolink:has_phenotype 1482455
#> 3                                biolink:expressed_in 1229165
#> 4                        biolink:actively_involved_in  632984
#> 5                              biolink:orthologous_to  551358
#> 6                                 biolink:subclass_of  544535
#> 7                                     biolink:enables  431854
#> 8                                  biolink:located_in  312598
#> 9                             biolink:participates_in  268142
#> 10                                 biolink:related_to  256225
#> 11                 biolink:acts_upstream_of_or_within  145303
#> 12                                  biolink:active_in  144810
#> 13                                    biolink:part_of   63174
#> 14                                     biolink:causes   15651
#> 15                     biolink:is_sequence_variant_of   13100
#> 16                                   biolink:model_of    9258
#> 17                    biolink:has_mode_of_inheritance    8205
#> 18             biolink:gene_associated_with_condition    8052
#> 19                           biolink:acts_upstream_of    7095
#> 20                             biolink:contributes_to    6332
#> 21      biolink:treats_or_applied_or_studied_to_treat    5335
#> 22                           biolink:colocalizes_with    2653
#> 23    biolink:associated_with_increased_likelihood_of    2208
#> 24                biolink:genetically_associated_with    2155
#> 25           biolink:acts_upstream_of_positive_effect     473
#> 26 biolink:acts_upstream_of_or_within_positive_effect     440
#> 27           biolink:acts_upstream_of_negative_effect     165
#> 28 biolink:acts_upstream_of_or_within_negative_effect     152
#> 
#> $total_nodes
#> [1] 1067161
#> 
#> $total_edges
#> [1] 8627308
#> 
#> $cats
#> $cats$`biolink:Entity`
#> [1] "biolink:Entity"
#> 
#> $cats$`biolink:NamedThing`
#> [1] "biolink:NamedThing"
#> 
#> $cats$`biolink:ThingWithTaxon`
#> [1] "biolink:ThingWithTaxon"
#> 
#> $cats$`biolink:BiologicalEntity`
#> [1] "biolink:BiologicalEntity"
#> 
#> $cats$`biolink:PhysicalEssenceOrOccurrent`
#> [1] "biolink:PhysicalEssenceOrOccurrent"
#> 
#> $cats$`biolink:PhysicalEssence`
#> [1] "biolink:PhysicalEssence"
#> 
#> $cats$`biolink:OntologyClass`
#> [1] "biolink:OntologyClass"
#> 
#> $cats$`biolink:GenomicEntity`
#> [1] "biolink:GenomicEntity"
#> 
#> $cats$`biolink:ChemicalEntityOrGeneOrGeneProduct`
#> [1] "biolink:ChemicalEntityOrGeneOrGeneProduct"
#> 
#> $cats$`biolink:MacromolecularMachineMixin`
#> [1] "biolink:MacromolecularMachineMixin"
#> 
#> $cats$`biolink:GeneOrGeneProduct`
#> [1] "biolink:GeneOrGeneProduct"
#> 
#> $cats$`biolink:Gene`
#> [1] "biolink:Gene"
#> 
#> $cats$`biolink:DiseaseOrPhenotypicFeature`
#> [1] "biolink:DiseaseOrPhenotypicFeature"
#> 
#> $cats$`biolink:PhenotypicFeature`
#> [1] "biolink:PhenotypicFeature"
#> 
#> $cats$`biolink:Genotype`
#> [1] "biolink:Genotype"
#> 
#> $cats$`biolink:SubjectOfInvestigation`
#> [1] "biolink:SubjectOfInvestigation"
#> 
#> $cats$`biolink:OrganismalEntity`
#> [1] "biolink:OrganismalEntity"
#> 
#> $cats$`biolink:Occurrent`
#> [1] "biolink:Occurrent"
#> 
#> $cats$`biolink:BiologicalProcessOrActivity`
#> [1] "biolink:BiologicalProcessOrActivity"
#> 
#> $cats$`biolink:AnatomicalEntity`
#> [1] "biolink:AnatomicalEntity"
#> 
#> $cats$`biolink:Disease`
#> [1] "biolink:Disease"
#> 
#> $cats$`biolink:BiologicalProcess`
#> [1] "biolink:BiologicalProcess"
#> 
#> $cats$`biolink:Cell`
#> [1] "biolink:Cell"
#> 
#> $cats$`biolink:GrossAnatomicalStructure`
#> [1] "biolink:GrossAnatomicalStructure"
#> 
#> $cats$`biolink:Pathway`
#> [1] "biolink:Pathway"
#> 
#> $cats$`biolink:SequenceVariant`
#> [1] "biolink:SequenceVariant"
#> 
#> $cats$`biolink:ChemicalEntityOrProteinOrPolypeptide`
#> [1] "biolink:ChemicalEntityOrProteinOrPolypeptide"
#> 
#> $cats$`biolink:CellularComponent`
#> [1] "biolink:CellularComponent"
#> 
#> $cats$`biolink:ChemicalOrDrugOrTreatment`
#> [1] "biolink:ChemicalOrDrugOrTreatment"
#> 
#> $cats$`biolink:ChemicalEntity`
#> [1] "biolink:ChemicalEntity"
#> 
#> $cats$`biolink:MolecularEntity`
#> [1] "biolink:MolecularEntity"
#> 
#> $cats$`biolink:MacromolecularComplex`
#> [1] "biolink:MacromolecularComplex"
#> 
#> $cats$`biolink:MolecularActivity`
#> [1] "biolink:MolecularActivity"
#> 
#> $cats$`biolink:CellularOrganism`
#> [1] "biolink:CellularOrganism"
#> 
#> $cats$`biolink:GeneProductMixin`
#> [1] "biolink:GeneProductMixin"
#> 
#> $cats$`biolink:Polypeptide`
#> [1] "biolink:Polypeptide"
#> 
#> $cats$`biolink:Protein`
#> [1] "biolink:Protein"
#> 
#> $cats$`biolink:Vertebrate`
#> [1] "biolink:Vertebrate"
#> 
#> $cats$`biolink:Virus`
#> [1] "biolink:Virus"
#> 
#> $cats$`biolink:BehavioralFeature`
#> [1] "biolink:BehavioralFeature"
#> 
#> $cats$`biolink:LifeStage`
#> [1] "biolink:LifeStage"
#> 
#> $cats$`biolink:PathologicalProcess`
#> [1] "biolink:PathologicalProcess"
#> 
#> $cats$`biolink:PathologicalEntityMixin`
#> [1] "biolink:PathologicalEntityMixin"
#> 
#> $cats$`biolink:ChemicalMixture`
#> [1] "biolink:ChemicalMixture"
#> 
#> $cats$`biolink:MolecularMixture`
#> [1] "biolink:MolecularMixture"
#> 
#> $cats$`biolink:Drug`
#> [1] "biolink:Drug"
#> 
#> $cats$`biolink:SmallMolecule`
#> [1] "biolink:SmallMolecule"
#> 
#> $cats$`biolink:Attribute`
#> [1] "biolink:Attribute"
#> 
#> $cats$`biolink:InformationContentEntity`
#> [1] "biolink:InformationContentEntity"
#> 
#> $cats$`biolink:ClinicalAttribute`
#> [1] "biolink:ClinicalAttribute"
#> 
#> $cats$`biolink:Onset`
#> [1] "biolink:Onset"
#> 
#> $cats$`biolink:ClinicalCourse`
#> [1] "biolink:ClinicalCourse"
#> 
#> $cats$`biolink:OrganismTaxon`
#> [1] "biolink:OrganismTaxon"
#> 
#> $cats$`biolink:Transcript`
#> [1] "biolink:Transcript"
#> 
#> $cats$`biolink:NucleicAcidEntity`
#> [1] "biolink:NucleicAcidEntity"
#> 
#> $cats$`biolink:EvidenceType`
#> [1] "biolink:EvidenceType"
#> 
#> $cats$`biolink:RNAProduct`
#> [1] "biolink:RNAProduct"
#> 
#> $cats$`biolink:Plant`
#> [1] "biolink:Plant"
#> 
#> $cats$`biolink:Publication`
#> [1] "biolink:Publication"
#> 
#> $cats$`biolink:Fungus`
#> [1] "biolink:Fungus"
#> 
#> $cats$`biolink:ExposureEvent`
#> [1] "biolink:ExposureEvent"
#> 
#> $cats$`biolink:PlanetaryEntity`
#> [1] "biolink:PlanetaryEntity"
#> 
#> $cats$`biolink:ProcessedMaterial`
#> [1] "biolink:ProcessedMaterial"
#> 
#> $cats$`biolink:Activity`
#> [1] "biolink:Activity"
#> 
#> $cats$`biolink:ActivityAndBehavior`
#> [1] "biolink:ActivityAndBehavior"
#> 
#> $cats$`biolink:Mammal`
#> [1] "biolink:Mammal"
#> 
#> $cats$`biolink:GeneGroupingMixin`
#> [1] "biolink:GeneGroupingMixin"
#> 
#> $cats$`biolink:BiologicalSex`
#> [1] "biolink:BiologicalSex"
#> 
#> $cats$`biolink:RegulatoryRegion`
#> [1] "biolink:RegulatoryRegion"
#> 
#> $cats$`biolink:ChemicalExposure`
#> [1] "biolink:ChemicalExposure"
#> 
#> $cats$`biolink:EnvironmentalFeature`
#> [1] "biolink:EnvironmentalFeature"
#> 
#> $cats$`biolink:AdministrativeEntity`
#> [1] "biolink:AdministrativeEntity"
#> 
#> $cats$`biolink:Agent`
#> [1] "biolink:Agent"
#> 
#> $cats$`biolink:Dataset`
#> [1] "biolink:Dataset"
#> 
#> $cats$`biolink:ConfidenceLevel`
#> [1] "biolink:ConfidenceLevel"
#> 
#> $cats$`biolink:Invertebrate`
#> [1] "biolink:Invertebrate"
#> 
#> $cats$`biolink:GeneticInheritance`
#> [1] "biolink:GeneticInheritance"
#> 
#> $cats$`biolink:Haplotype`
#> [1] "biolink:Haplotype"
#> 
#> $cats$`biolink:ProteinFamily`
#> [1] "biolink:ProteinFamily"
#> 
#> $cats$`biolink:PopulationOfIndividualOrganisms`
#> [1] "biolink:PopulationOfIndividualOrganisms"
#> 
#> $cats$`biolink:NoncodingRNAProduct`
#> [1] "biolink:NoncodingRNAProduct"
#> 
#> $cats$`biolink:DrugExposure`
#> [1] "biolink:DrugExposure"
#> 
#> $cats$`biolink:WebPage`
#> [1] "biolink:WebPage"
#> 
#> $cats$`biolink:DatasetDistribution`
#> [1] "biolink:DatasetDistribution"
#> 
#> $cats$`biolink:IndividualOrganism`
#> [1] "biolink:IndividualOrganism"
#> 
#> $cats$`biolink:Patent`
#> [1] "biolink:Patent"
#> 
#> $cats$`biolink:CellLine`
#> [1] "biolink:CellLine"
#> 
#> $cats$`biolink:EnvironmentalProcess`
#> [1] "biolink:EnvironmentalProcess"
#> 
#> $cats$`biolink:Zygosity`
#> [1] "biolink:Zygosity"
#> 
#> $cats$`biolink:ReagentTargetedGene`
#> [1] "biolink:ReagentTargetedGene"
#> 
#> $cats$`biolink:Bacterium`
#> [1] "biolink:Bacterium"
#> 
#> $cats$`biolink:ProteinDomain`
#> [1] "biolink:ProteinDomain"
#> 
#> $cats$`biolink:StudyVariable`
#> [1] "biolink:StudyVariable"
#> 
#> $cats$`biolink:Human`
#> [1] "biolink:Human"
#> 
#> $cats$`biolink:Event`
#> [1] "biolink:Event"
#> 
#> $cats$`biolink:Study`
#> [1] "biolink:Study"
#> 
#> $cats$`biolink:MaterialSample`
#> [1] "biolink:MaterialSample"
#> 
#> $cats$`biolink:PhysicalEntity`
#> [1] "biolink:PhysicalEntity"
#> 
#> $cats$`biolink:Treatment`
#> [1] "biolink:Treatment"
#> 
#> $cats$`biolink:PhenotypicSex`
#> [1] "biolink:PhenotypicSex"
#> 
#> $cats$`biolink:GenotypicSex`
#> [1] "biolink:GenotypicSex"
#> 
#> $cats$`biolink:Exon`
#> [1] "biolink:Exon"
#> 
#> $cats$`biolink:TranscriptionFactorBindingSite`
#> [1] "biolink:TranscriptionFactorBindingSite"
#> 
#> $cats$`biolink:MicroRNA`
#> [1] "biolink:MicroRNA"
#> 
#> $cats$`biolink:SiRNA`
#> [1] "biolink:SiRNA"
#> 
#> $cats$`biolink:Genome`
#> [1] "biolink:Genome"
#> 
#> $cats$`biolink:Snv`
#> [1] "biolink:Snv"
#> 
#> $cats$`biolink:AccessibleDnaRegion`
#> [1] "biolink:AccessibleDnaRegion"
#> 
#> $cats$`biolink:DiagnosticAid`
#> [1] "biolink:DiagnosticAid"
#> 
#> 
#> $preds
#> $preds$`biolink:interacts_with`
#> [1] "biolink:interacts_with"
#> 
#> $preds$`biolink:has_phenotype`
#> [1] "biolink:has_phenotype"
#> 
#> $preds$`biolink:expressed_in`
#> [1] "biolink:expressed_in"
#> 
#> $preds$`biolink:actively_involved_in`
#> [1] "biolink:actively_involved_in"
#> 
#> $preds$`biolink:orthologous_to`
#> [1] "biolink:orthologous_to"
#> 
#> $preds$`biolink:subclass_of`
#> [1] "biolink:subclass_of"
#> 
#> $preds$`biolink:enables`
#> [1] "biolink:enables"
#> 
#> $preds$`biolink:located_in`
#> [1] "biolink:located_in"
#> 
#> $preds$`biolink:participates_in`
#> [1] "biolink:participates_in"
#> 
#> $preds$`biolink:related_to`
#> [1] "biolink:related_to"
#> 
#> $preds$`biolink:acts_upstream_of_or_within`
#> [1] "biolink:acts_upstream_of_or_within"
#> 
#> $preds$`biolink:active_in`
#> [1] "biolink:active_in"
#> 
#> $preds$`biolink:part_of`
#> [1] "biolink:part_of"
#> 
#> $preds$`biolink:causes`
#> [1] "biolink:causes"
#> 
#> $preds$`biolink:is_sequence_variant_of`
#> [1] "biolink:is_sequence_variant_of"
#> 
#> $preds$`biolink:model_of`
#> [1] "biolink:model_of"
#> 
#> $preds$`biolink:has_mode_of_inheritance`
#> [1] "biolink:has_mode_of_inheritance"
#> 
#> $preds$`biolink:gene_associated_with_condition`
#> [1] "biolink:gene_associated_with_condition"
#> 
#> $preds$`biolink:acts_upstream_of`
#> [1] "biolink:acts_upstream_of"
#> 
#> $preds$`biolink:contributes_to`
#> [1] "biolink:contributes_to"
#> 
#> $preds$`biolink:treats_or_applied_or_studied_to_treat`
#> [1] "biolink:treats_or_applied_or_studied_to_treat"
#> 
#> $preds$`biolink:colocalizes_with`
#> [1] "biolink:colocalizes_with"
#> 
#> $preds$`biolink:associated_with_increased_likelihood_of`
#> [1] "biolink:associated_with_increased_likelihood_of"
#> 
#> $preds$`biolink:genetically_associated_with`
#> [1] "biolink:genetically_associated_with"
#> 
#> $preds$`biolink:acts_upstream_of_positive_effect`
#> [1] "biolink:acts_upstream_of_positive_effect"
#> 
#> $preds$`biolink:acts_upstream_of_or_within_positive_effect`
#> [1] "biolink:acts_upstream_of_or_within_positive_effect"
#> 
#> $preds$`biolink:acts_upstream_of_negative_effect`
#> [1] "biolink:acts_upstream_of_negative_effect"
#> 
#> $preds$`biolink:acts_upstream_of_or_within_negative_effect`
#> [1] "biolink:acts_upstream_of_or_within_negative_effect"
#> 
#>