Summarize contents of a Neo4j KG engine

Given a Neo4j based KG engine, provides summary information in the form of node counts, category counts across nodes, relationship type counts, and available properties. General information about the graph is printed to the console, and a list of dataframes with this information is returned invisibly. Also returned are cats, preds, and props entries, containing lists of available categories/predicates/properties for convenient auto-completion in RStudio.

# S3 method for class 'neo4j_engine'
summary(object, ..., quiet = FALSE)

Arguments

object

A neo4j_engine object

...

Other parameters (not used)

quiet

Logical, whether to suppress printing of the summary

Value

A list of dataframes and named lists

Examples

# prints a readable summary and returns a list of dataframes
stats <- monarch_engine() |> summary()
#> Trying to connect to https://neo4j.monarchinitiative.org
#> Connected to https://neo4j.monarchinitiative.org
#> 
#> A Neo4j-backed knowledge graph engine.
#> Gathering statistics, please wait...
#> Total nodes:  1346185 
#> Total edges:  14621271 
#> 
#> Node category counts:
#>                   category  count
#>               biolink:Gene 581682
#>    biolink:SequenceVariant 209865
#>  biolink:PhenotypicFeature 169316
#>           biolink:Genotype 137639
#>   biolink:AnatomicalEntity  66029
#>  biolink:BiologicalProcess  30769
#>            biolink:Disease  30384
#>         biolink:NamedThing  27656
#>            biolink:Protein  25098
#>            biolink:Pathway  22254
#>     biolink:ChemicalEntity  22231
#>  biolink:MolecularActivity  12723
#>  biolink:CellularComponent   4538
#>               biolink:Cell   3171
#>      biolink:OrganismTaxon   1987
#>    biolink:MolecularEntity    529
#>          biolink:LifeStage    314
#> 
#> Edge type counts:
#>                                           predicate   count
#>                              biolink:interacts_with 2797006
#>                                biolink:expressed_in 2379934
#>                               biolink:has_phenotype 2077069
#>                              biolink:orthologous_to 1506263
#>                                     biolink:enables 1239419
#>                        biolink:actively_involved_in 1103592
#>                                  biolink:located_in  994683
#>                                 biolink:subclass_of  612920
#>                                  biolink:related_to  520067
#>                             biolink:participates_in  329190
#>                  biolink:acts_upstream_of_or_within  295324
#>                                biolink:is_active_in  188945
#>                      biolink:is_sequence_variant_of  181003
#>                        biolink:has_sequence_variant  157143
#>                                     biolink:part_of  129977
#>       biolink:treats_or_applied_or_studied_to_treat   18934
#>                                      biolink:causes   18203
#>                            biolink:acts_upstream_of   17896
#>                                    biolink:model_of   10069
#>                              biolink:contributes_to    9837
#>                     biolink:has_mode_of_inheritance    8829
#>              biolink:gene_associated_with_condition    8089
#>     biolink:associated_with_increased_likelihood_of    4394
#>                            biolink:colocalizes_with    4251
#>                 biolink:genetically_associated_with    3484
#>                        biolink:disease_has_location    1020
#>            biolink:acts_upstream_of_positive_effect     992
#>  biolink:acts_upstream_of_or_within_positive_effect     826
#>                       biolink:ameliorates_condition     786
#>            biolink:acts_upstream_of_negative_effect     336
#>  biolink:acts_upstream_of_or_within_negative_effect     279
#>                             biolink:has_participant     269
#>                  biolink:preventative_for_condition     145
#>                                    biolink:disrupts      89
#>                                   biolink:caused_by       5
#>                          biolink:contraindicated_in       3
#> 
#> Available node and edge properties:
#>  [1] "category"                         "name"                            
#>  [3] "provided_by"                      "in_taxon"                        
#>  [5] "in_taxon_label"                   "namespace"                       
#>  [7] "id"                               "description"                     
#>  [9] "synonym"                          "exact_synonyms"                  
#> [11] "iri"                              "deprecated"                      
#> [13] "full_name"                        "symbol"                          
#> [15] "xref"                             "type"                            
#> [17] "narrow_synonyms"                  "broad_synonyms"                  
#> [19] "related_synonyms"                 "subsets"                         
#> [21] "has_gene"                         "predicate"                       
#> [23] "agent_type"                       "aggregator_knowledge_source"     
#> [25] "knowledge_level"                  "primary_knowledge_source"        
#> [27] "publications"                     "subject"                         
#> [29] "object"                           "original_subject"                
#> [31] "original_object"                  "has_evidence"                    
#> [33] "disease_context_qualifier"        "has_count"                       
#> [35] "has_percentage"                   "has_quotient"                    
#> [37] "has_total"                        "frequency_qualifier"             
#> [39] "qualifiers"                       "stage_qualifier"                 
#> [41] "original_predicate"               "subject_specialization_qualifier"
#> [43] "qualifier"                        "object_specialization_qualifier" 
#> [45] "negated"                          "onset_qualifier"                 
#> [47] "sex_qualifier"                    "species_context_qualifier"       
#> 
#> 
#> For more information about Biolink node (Class) and edge (Association) properties, see https://biolink.github.io/biolink-model/.
print(stats)
#> $node_summary
#>                     category  count
#> 1               biolink:Gene 581682
#> 2    biolink:SequenceVariant 209865
#> 3  biolink:PhenotypicFeature 169316
#> 4           biolink:Genotype 137639
#> 5   biolink:AnatomicalEntity  66029
#> 6  biolink:BiologicalProcess  30769
#> 7            biolink:Disease  30384
#> 8         biolink:NamedThing  27656
#> 9            biolink:Protein  25098
#> 10           biolink:Pathway  22254
#> 11    biolink:ChemicalEntity  22231
#> 12 biolink:MolecularActivity  12723
#> 13 biolink:CellularComponent   4538
#> 14              biolink:Cell   3171
#> 15     biolink:OrganismTaxon   1987
#> 16   biolink:MolecularEntity    529
#> 17         biolink:LifeStage    314
#> 
#> $edge_summary
#>                                             predicate   count
#> 1                              biolink:interacts_with 2797006
#> 2                                biolink:expressed_in 2379934
#> 3                               biolink:has_phenotype 2077069
#> 4                              biolink:orthologous_to 1506263
#> 5                                     biolink:enables 1239419
#> 6                        biolink:actively_involved_in 1103592
#> 7                                  biolink:located_in  994683
#> 8                                 biolink:subclass_of  612920
#> 9                                  biolink:related_to  520067
#> 10                            biolink:participates_in  329190
#> 11                 biolink:acts_upstream_of_or_within  295324
#> 12                               biolink:is_active_in  188945
#> 13                     biolink:is_sequence_variant_of  181003
#> 14                       biolink:has_sequence_variant  157143
#> 15                                    biolink:part_of  129977
#> 16      biolink:treats_or_applied_or_studied_to_treat   18934
#> 17                                     biolink:causes   18203
#> 18                           biolink:acts_upstream_of   17896
#> 19                                   biolink:model_of   10069
#> 20                             biolink:contributes_to    9837
#> 21                    biolink:has_mode_of_inheritance    8829
#> 22             biolink:gene_associated_with_condition    8089
#> 23    biolink:associated_with_increased_likelihood_of    4394
#> 24                           biolink:colocalizes_with    4251
#> 25                biolink:genetically_associated_with    3484
#> 26                       biolink:disease_has_location    1020
#> 27           biolink:acts_upstream_of_positive_effect     992
#> 28 biolink:acts_upstream_of_or_within_positive_effect     826
#> 29                      biolink:ameliorates_condition     786
#> 30           biolink:acts_upstream_of_negative_effect     336
#> 31 biolink:acts_upstream_of_or_within_negative_effect     279
#> 32                            biolink:has_participant     269
#> 33                 biolink:preventative_for_condition     145
#> 34                                   biolink:disrupts      89
#> 35                                  biolink:caused_by       5
#> 36                         biolink:contraindicated_in       3
#> 
#> $total_nodes
#> [1] 1346185
#> 
#> $total_edges
#> [1] 14621271
#> 
#> $cats
#> $cats$`biolink:Gene`
#> [1] "biolink:Gene"
#> 
#> $cats$`biolink:SequenceVariant`
#> [1] "biolink:SequenceVariant"
#> 
#> $cats$`biolink:PhenotypicFeature`
#> [1] "biolink:PhenotypicFeature"
#> 
#> $cats$`biolink:Genotype`
#> [1] "biolink:Genotype"
#> 
#> $cats$`biolink:AnatomicalEntity`
#> [1] "biolink:AnatomicalEntity"
#> 
#> $cats$`biolink:BiologicalProcess`
#> [1] "biolink:BiologicalProcess"
#> 
#> $cats$`biolink:Disease`
#> [1] "biolink:Disease"
#> 
#> $cats$`biolink:NamedThing`
#> [1] "biolink:NamedThing"
#> 
#> $cats$`biolink:Protein`
#> [1] "biolink:Protein"
#> 
#> $cats$`biolink:Pathway`
#> [1] "biolink:Pathway"
#> 
#> $cats$`biolink:ChemicalEntity`
#> [1] "biolink:ChemicalEntity"
#> 
#> $cats$`biolink:MolecularActivity`
#> [1] "biolink:MolecularActivity"
#> 
#> $cats$`biolink:CellularComponent`
#> [1] "biolink:CellularComponent"
#> 
#> $cats$`biolink:Cell`
#> [1] "biolink:Cell"
#> 
#> $cats$`biolink:OrganismTaxon`
#> [1] "biolink:OrganismTaxon"
#> 
#> $cats$`biolink:MolecularEntity`
#> [1] "biolink:MolecularEntity"
#> 
#> $cats$`biolink:LifeStage`
#> [1] "biolink:LifeStage"
#> 
#> 
#> $preds
#> $preds$`biolink:interacts_with`
#> [1] "biolink:interacts_with"
#> 
#> $preds$`biolink:expressed_in`
#> [1] "biolink:expressed_in"
#> 
#> $preds$`biolink:has_phenotype`
#> [1] "biolink:has_phenotype"
#> 
#> $preds$`biolink:orthologous_to`
#> [1] "biolink:orthologous_to"
#> 
#> $preds$`biolink:enables`
#> [1] "biolink:enables"
#> 
#> $preds$`biolink:actively_involved_in`
#> [1] "biolink:actively_involved_in"
#> 
#> $preds$`biolink:located_in`
#> [1] "biolink:located_in"
#> 
#> $preds$`biolink:subclass_of`
#> [1] "biolink:subclass_of"
#> 
#> $preds$`biolink:related_to`
#> [1] "biolink:related_to"
#> 
#> $preds$`biolink:participates_in`
#> [1] "biolink:participates_in"
#> 
#> $preds$`biolink:acts_upstream_of_or_within`
#> [1] "biolink:acts_upstream_of_or_within"
#> 
#> $preds$`biolink:is_active_in`
#> [1] "biolink:is_active_in"
#> 
#> $preds$`biolink:is_sequence_variant_of`
#> [1] "biolink:is_sequence_variant_of"
#> 
#> $preds$`biolink:has_sequence_variant`
#> [1] "biolink:has_sequence_variant"
#> 
#> $preds$`biolink:part_of`
#> [1] "biolink:part_of"
#> 
#> $preds$`biolink:treats_or_applied_or_studied_to_treat`
#> [1] "biolink:treats_or_applied_or_studied_to_treat"
#> 
#> $preds$`biolink:causes`
#> [1] "biolink:causes"
#> 
#> $preds$`biolink:acts_upstream_of`
#> [1] "biolink:acts_upstream_of"
#> 
#> $preds$`biolink:model_of`
#> [1] "biolink:model_of"
#> 
#> $preds$`biolink:contributes_to`
#> [1] "biolink:contributes_to"
#> 
#> $preds$`biolink:has_mode_of_inheritance`
#> [1] "biolink:has_mode_of_inheritance"
#> 
#> $preds$`biolink:gene_associated_with_condition`
#> [1] "biolink:gene_associated_with_condition"
#> 
#> $preds$`biolink:associated_with_increased_likelihood_of`
#> [1] "biolink:associated_with_increased_likelihood_of"
#> 
#> $preds$`biolink:colocalizes_with`
#> [1] "biolink:colocalizes_with"
#> 
#> $preds$`biolink:genetically_associated_with`
#> [1] "biolink:genetically_associated_with"
#> 
#> $preds$`biolink:disease_has_location`
#> [1] "biolink:disease_has_location"
#> 
#> $preds$`biolink:acts_upstream_of_positive_effect`
#> [1] "biolink:acts_upstream_of_positive_effect"
#> 
#> $preds$`biolink:acts_upstream_of_or_within_positive_effect`
#> [1] "biolink:acts_upstream_of_or_within_positive_effect"
#> 
#> $preds$`biolink:ameliorates_condition`
#> [1] "biolink:ameliorates_condition"
#> 
#> $preds$`biolink:acts_upstream_of_negative_effect`
#> [1] "biolink:acts_upstream_of_negative_effect"
#> 
#> $preds$`biolink:acts_upstream_of_or_within_negative_effect`
#> [1] "biolink:acts_upstream_of_or_within_negative_effect"
#> 
#> $preds$`biolink:has_participant`
#> [1] "biolink:has_participant"
#> 
#> $preds$`biolink:preventative_for_condition`
#> [1] "biolink:preventative_for_condition"
#> 
#> $preds$`biolink:disrupts`
#> [1] "biolink:disrupts"
#> 
#> $preds$`biolink:caused_by`
#> [1] "biolink:caused_by"
#> 
#> $preds$`biolink:contraindicated_in`
#> [1] "biolink:contraindicated_in"
#> 
#> 
#> $props
#> $props$category
#> [1] "category"
#> 
#> $props$name
#> [1] "name"
#> 
#> $props$provided_by
#> [1] "provided_by"
#> 
#> $props$in_taxon
#> [1] "in_taxon"
#> 
#> $props$in_taxon_label
#> [1] "in_taxon_label"
#> 
#> $props$namespace
#> [1] "namespace"
#> 
#> $props$id
#> [1] "id"
#> 
#> $props$description
#> [1] "description"
#> 
#> $props$synonym
#> [1] "synonym"
#> 
#> $props$exact_synonyms
#> [1] "exact_synonyms"
#> 
#> $props$iri
#> [1] "iri"
#> 
#> $props$deprecated
#> [1] "deprecated"
#> 
#> $props$full_name
#> [1] "full_name"
#> 
#> $props$symbol
#> [1] "symbol"
#> 
#> $props$xref
#> [1] "xref"
#> 
#> $props$type
#> [1] "type"
#> 
#> $props$narrow_synonyms
#> [1] "narrow_synonyms"
#> 
#> $props$broad_synonyms
#> [1] "broad_synonyms"
#> 
#> $props$related_synonyms
#> [1] "related_synonyms"
#> 
#> $props$subsets
#> [1] "subsets"
#> 
#> $props$has_gene
#> [1] "has_gene"
#> 
#> $props$predicate
#> [1] "predicate"
#> 
#> $props$agent_type
#> [1] "agent_type"
#> 
#> $props$aggregator_knowledge_source
#> [1] "aggregator_knowledge_source"
#> 
#> $props$knowledge_level
#> [1] "knowledge_level"
#> 
#> $props$primary_knowledge_source
#> [1] "primary_knowledge_source"
#> 
#> $props$publications
#> [1] "publications"
#> 
#> $props$subject
#> [1] "subject"
#> 
#> $props$object
#> [1] "object"
#> 
#> $props$original_subject
#> [1] "original_subject"
#> 
#> $props$original_object
#> [1] "original_object"
#> 
#> $props$has_evidence
#> [1] "has_evidence"
#> 
#> $props$disease_context_qualifier
#> [1] "disease_context_qualifier"
#> 
#> $props$has_count
#> [1] "has_count"
#> 
#> $props$has_percentage
#> [1] "has_percentage"
#> 
#> $props$has_quotient
#> [1] "has_quotient"
#> 
#> $props$has_total
#> [1] "has_total"
#> 
#> $props$frequency_qualifier
#> [1] "frequency_qualifier"
#> 
#> $props$qualifiers
#> [1] "qualifiers"
#> 
#> $props$stage_qualifier
#> [1] "stage_qualifier"
#> 
#> $props$original_predicate
#> [1] "original_predicate"
#> 
#> $props$subject_specialization_qualifier
#> [1] "subject_specialization_qualifier"
#> 
#> $props$qualifier
#> [1] "qualifier"
#> 
#> $props$object_specialization_qualifier
#> [1] "object_specialization_qualifier"
#> 
#> $props$negated
#> [1] "negated"
#> 
#> $props$onset_qualifier
#> [1] "onset_qualifier"
#> 
#> $props$sex_qualifier
#> [1] "sex_qualifier"
#> 
#> $props$species_context_qualifier
#> [1] "species_context_qualifier"
#> 
#>